@@ -1,7 +1,8 @@
 package fr.ralala.worktime.ui.adapters;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.List;
 import java.util.Locale;
@@ -17,30 +18,32 @@ import fr.ralala.worktime.models.ProfileEntry;
  * </p>
  *
  * @author Keidan
- *         <p>
- *******************************************************************************
+ * <p>
+ * ******************************************************************************
  */
-public class ProfilesEntriesArrayAdapter  extends EntriesArrayAdapter<ProfileEntry> {
+public class ProfilesEntriesArrayAdapter extends EntriesArrayAdapter<ProfileEntry> {
 
-  private Context mContext;
+  private final Context mContext;
 
   /**
    * Creates the array adapter.
-   * @param recyclerView The owner object.
-   * @param context The Android context.
+   *
+   * @param recyclerView  The owner object.
+   * @param context       The Android context.
    * @param rowResourceId The resource id of the container.
-   * @param objects The objects list.
+   * @param objects       The objects list.
    */
   public ProfilesEntriesArrayAdapter(RecyclerView recyclerView, final Context context, final int rowResourceId,
-                                           final List<ProfileEntry> objects) {
+                                     final List<ProfileEntry> objects) {
     super(recyclerView, rowResourceId, objects);
     mContext = context;
   }
 
   /**
    * Called on Binding the view holder.
+   *
    * @param viewHolder The view holder.
-   * @param pe The associated item.
+   * @param pe         The associated item.
    */
   @Override
   public void onBindViewHolderEntry(EntriesArrayAdapter.ViewHolder viewHolder, ProfileEntry pe) {
@@ -48,30 +51,30 @@ public class ProfilesEntriesArrayAdapter  extends EntriesArrayAdapter<ProfileEnt
       if (viewHolder.name != null)
         viewHolder.name.setText(pe.getName());
       if (viewHolder.info != null) {
-        MainApplication app = MainApplication.getInstance();
+        MainApplication app = (MainApplication) mContext.getApplicationContext();
         viewHolder.info.setText(!app.isHideWage() ?
-            String.format(Locale.US,
-                "%s %s %s -> %s %s %.02f%s/h %s %s %s",
-                pe.getStartMorning().timeString(),
-                mContext.getString(R.string.at),
-                pe.getEndAfternoon().timeString(),
-                pe.getWorkTime().timeString(),
-                mContext.getString(R.string.text_for),
-                pe.getAmountByHour(),
-                MainApplication.getInstance().getCurrency(),
-                mContext.getString(R.string.and),
-                pe.getOverTime().timeString(),
-                mContext.getString(R.string.more))
-            :
-            String.format(Locale.US,
-                "%s %s %s -> %s %s %s %s",
-                pe.getStartMorning().timeString(),
-                mContext.getString(R.string.at),
-                pe.getEndAfternoon().timeString(),
-                pe.getWorkTime().timeString(),
-                mContext.getString(R.string.and),
-                pe.getOverTime().timeString(),
-                mContext.getString(R.string.more)));
+          String.format(Locale.US,
+            "%s %s %s -> %s %s %.02f%s/h %s %s %s",
+            pe.getStartMorning().timeString(),
+            mContext.getString(R.string.at),
+            pe.getEndAfternoon().timeString(),
+            pe.getWorkTime().timeString(),
+            mContext.getString(R.string.text_for),
+            pe.getAmountByHour(),
+            app.getCurrency(),
+            mContext.getString(R.string.and),
+            pe.getOverTime().timeString(),
+            mContext.getString(R.string.more))
+          :
+          String.format(Locale.US,
+            "%s %s %s -> %s %s %s %s",
+            pe.getStartMorning().timeString(),
+            mContext.getString(R.string.at),
+            pe.getEndAfternoon().timeString(),
+            pe.getWorkTime().timeString(),
+            mContext.getString(R.string.and),
+            pe.getOverTime().timeString(),
+            mContext.getString(R.string.more)));
       }
     }
   }
