@@ -1,77 +1,78 @@
 package fr.ralala.worktime.ui.fragments;
 
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.google.android.material.navigation.NavigationView;
 
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Manage the application fragments
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class AppFragmentsFactory {
-  public static final int IDX_WORK_TIME = 1;
-  public static final int IDX_QUICK_ACCESS = 0;
-  public static final int IDX_PROFILE = 2;
-  public static final int IDX_PUBLIC_HOLIDAY = 3;
-  public static final int IDX_EXPORT = 4;
-  public static final int IDX_STATISTICS = 6;
-  private static final int IDX_EXIT = 7;
-  private Fragment mQuickAccessFragment = null;
+  public static final int IDX_WORK_TIME = 0;
+  public static final int IDX_PROFILE = 1;
+  public static final int IDX_PUBLIC_HOLIDAY = 2;
+  public static final int IDX_EXPORT = 3;
+  public static final int IDX_STATISTICS = 4;
+  private static final int IDX_EXIT = 6;
   private Fragment mProfileFragment = null;
   private Fragment mPublicHolidaysFragment = null;
   private Fragment mExportFragment = null;
   private Fragment mWorkTimeFragment = null;
   private Fragment mStatisticsFragment = null;
   private Fragment mCurrentFragment;
-  private MainApplication mApp;
-  private NavigationView mNavigationView;
+  private final MainApplication mApp;
+  private final NavigationView mNavigationView;
 
   /**
    * Creates the fragments factory.
-   * @param app The application context.
+   *
+   * @param app            The application context.
    * @param navigationView The application navigation view.
    */
   public AppFragmentsFactory(final MainApplication app, final NavigationView navigationView) {
     mApp = app;
     mNavigationView = navigationView;
-    if(mQuickAccessFragment == null)
-      mQuickAccessFragment = new QuickAccessFragment();
-    if(mProfileFragment == null)
+    if (mProfileFragment == null)
       mProfileFragment = new ProfileFragment();
-    if(mPublicHolidaysFragment == null)
+    if (mPublicHolidaysFragment == null)
       mPublicHolidaysFragment = new PublicHolidaysFragment();
-    if(mExportFragment == null)
+    if (mExportFragment == null)
       mExportFragment = new ExportFragment();
-    if(mStatisticsFragment == null)
+    if (mStatisticsFragment == null)
       mStatisticsFragment = new StatisticsFragment();
-    if(mWorkTimeFragment == null)
+    if (mWorkTimeFragment == null)
       mWorkTimeFragment = new WorkTimeFragment();
     mCurrentFragment = mWorkTimeFragment;
   }
 
   /**
    * Tests whether the current fragment requires the progress dialog.
+   *
    * @return boolean
    */
   public boolean isRequiresProgress() {
     return mCurrentFragment != null &&
-        (mCurrentFragment.equals(mWorkTimeFragment) ||
-          mCurrentFragment.equals(mProfileFragment) ||
-            mCurrentFragment.equals(mPublicHolidaysFragment) ||
-            mCurrentFragment.equals(mExportFragment));
+      (mCurrentFragment.equals(mWorkTimeFragment) ||
+        mCurrentFragment.equals(mProfileFragment) ||
+        mCurrentFragment.equals(mPublicHolidaysFragment) ||
+        mCurrentFragment.equals(mExportFragment));
   }
 
   /**
    * Returns the default home index.
+   *
    * @return int
    */
   public int getDefaultHomeIndex() {
@@ -81,31 +82,31 @@ public class AppFragmentsFactory {
 
   /**
    * Returns the current fragment ID (android).
+   *
    * @return int
    */
   public int getCurrentFragmentId() {
-    if(mCurrentFragment == null)
+    if (mCurrentFragment == null)
       return getDefaultHomeId();
-    else if(mCurrentFragment.equals(mProfileFragment))
+    else if (mCurrentFragment.equals(mProfileFragment))
       return (R.id.nav_profile);
-    else if(mCurrentFragment.equals(mPublicHolidaysFragment))
+    else if (mCurrentFragment.equals(mPublicHolidaysFragment))
       return (R.id.nav_public_holidays);
-    else if(mCurrentFragment.equals(mExportFragment))
+    else if (mCurrentFragment.equals(mExportFragment))
       return (R.id.nav_export);
-    else if(mCurrentFragment.equals(mStatisticsFragment))
+    else if (mCurrentFragment.equals(mStatisticsFragment))
       return (R.id.nav_statistics);
-    else if(mCurrentFragment.equals(mWorkTimeFragment))
-      return (R.id.nav_worktime);
     else
-      return (R.id.nav_quickaccess);
+      return (R.id.nav_worktime);
   }
 
   /**
    * Returns the default home ID (android).
+   *
    * @return int
    */
   public int getDefaultHomeId() {
-    switch(mApp.getDefaultHome()) {
+    switch (mApp.getDefaultHome()) {
       case IDX_PROFILE:
         return (R.id.nav_profile);
       case IDX_PUBLIC_HOLIDAY:
@@ -115,21 +116,18 @@ public class AppFragmentsFactory {
       case IDX_STATISTICS:
         return (R.id.nav_statistics);
       case IDX_WORK_TIME:
-        return (R.id.nav_worktime);
-      case IDX_QUICK_ACCESS:
       default:
-        return (R.id.nav_quickaccess);
+        return (R.id.nav_worktime);
     }
   }
 
   /**
    * Returns the default home fragment.
+   *
    * @return Fragment
    */
   private Fragment getDefaultHomeView() {
-    switch(mApp.getDefaultHome()) {
-      case IDX_QUICK_ACCESS:
-        return mQuickAccessFragment;
+    switch (mApp.getDefaultHome()) {
       case IDX_PROFILE:
         return mProfileFragment;
       case IDX_PUBLIC_HOLIDAY:
@@ -146,42 +144,43 @@ public class AppFragmentsFactory {
 
   /**
    * Called when the backPressed event is caught in the main activity.
+   *
    * @return true if the backPressed is consumed by a fragment.
    */
   public boolean consumeBackPressed() {
     Fragment fragment = mCurrentFragment;
-    return fragment != null && StatisticsFragment.class.isInstance(fragment) && ((StatisticsFragment)fragment).consumeBackPressed();
+    return fragment instanceof StatisticsFragment && ((StatisticsFragment) fragment).consumeBackPressed();
   }
 
   /**
    * Called when the activity is resumed.
+   *
    * @param aca The main activity.
    */
   public void onResume(final AppCompatActivity aca) {
     mNavigationView.getMenu().getItem(IDX_EXIT).setVisible(!mApp.isHideExitButton());
     Fragment fragment = mCurrentFragment;
-    if(fragment != null) {
+    if (fragment != null) {
       final FragmentTransaction ft = aca.getSupportFragmentManager().beginTransaction();
       ft.setCustomAnimations(R.anim.fadein, R.anim.fadeout);
       ft.replace(R.id.content_frame, fragment);
       ft.commit();
-      if(QuickAccessFragment.class.isInstance(fragment))
-        mNavigationView.getMenu().getItem(IDX_QUICK_ACCESS).setChecked(true); /* select quick access title */
-      else if(ProfileFragment.class.isInstance(fragment))
+      if (fragment instanceof ProfileFragment)
         mNavigationView.getMenu().getItem(IDX_PROFILE).setChecked(true); /* select profile title */
-      else if(PublicHolidaysFragment.class.isInstance(fragment))
+      else if (fragment instanceof PublicHolidaysFragment)
         mNavigationView.getMenu().getItem(IDX_PUBLIC_HOLIDAY).setChecked(true); /* select public holidays title */
-      else if(ExportFragment.class.isInstance(fragment))
+      else if (fragment instanceof ExportFragment)
         mNavigationView.getMenu().getItem(IDX_EXPORT).setChecked(true); /* select export title */
-      else if(WorkTimeFragment.class.isInstance(fragment))
+      else if (fragment instanceof WorkTimeFragment)
         mNavigationView.getMenu().getItem(IDX_WORK_TIME).setChecked(true); /* select work title */
-      else if(StatisticsFragment.class.isInstance(fragment))
+      else if (fragment instanceof StatisticsFragment)
         mNavigationView.getMenu().getItem(IDX_STATISTICS).setChecked(true); /* select charts title */
     }
   }
 
   /**
    * Return the current fragment.
+   *
    * @return Fragment.
    */
   public Fragment getCurrentFragment() {
@@ -190,13 +189,11 @@ public class AppFragmentsFactory {
 
   /**
    * Changes the current fragment based on its index in the navigation view.
+   *
    * @param idx The index.
    */
   public void setCurrentToFragment(int idx) {
     switch (idx) {
-      case IDX_QUICK_ACCESS:
-        mCurrentFragment = mQuickAccessFragment;
-        break;
       case IDX_PROFILE:
         mCurrentFragment = mProfileFragment;
         break;
