@@ -5,17 +5,17 @@ import android.content.SharedPreferences;
 
 import com.dropbox.core.DbxRequestConfig;
 import com.dropbox.core.android.Auth;
-import com.dropbox.core.http.OkHttp3Requestor;
 import com.dropbox.core.v2.DbxClientV2;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Helper functions
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 @SuppressWarnings("WeakerAccess")
 public class DropboxHelper {
@@ -24,27 +24,30 @@ public class DropboxHelper {
   private DbxClientV2 mDbxClient;
   private static DropboxHelper mHelper = null;
 
-  private DropboxHelper() { }
+  private DropboxHelper() {
+  }
 
   /**
    * Returns the helper instance.
+   *
    * @return DropboxHelper
    */
   static DropboxHelper helper() {
-    if(mHelper == null) mHelper = new DropboxHelper();
+    if (mHelper == null) mHelper = new DropboxHelper();
     return mHelper;
   }
 
   /**
    * Established the connection with dropbox
-   * @param ctx The Android context.
+   *
+   * @param ctx    The Android context.
    * @param appkey The application key.
    * @return boolean
    */
   boolean connect(final Context ctx, final String appkey) {
     Context c = ctx.getApplicationContext();
     mHelper.loadToken(c);
-    if(!mHelper.hasToken(c)) {
+    if (!mHelper.hasToken(c)) {
       Auth.startOAuth2Authentication(c, appkey);
       mHelper.loadToken(c);
       return mHelper.hasToken(c);
@@ -54,6 +57,7 @@ public class DropboxHelper {
 
   /**
    * Tests if the configuration has the required token.
+   *
    * @param ctx The Android context.
    * @return boolean
    */
@@ -66,6 +70,7 @@ public class DropboxHelper {
 
   /**
    * Loads the dropbox token.
+   *
    * @param ctx The Android context.
    */
   private void loadToken(final Context ctx) {
@@ -85,11 +90,13 @@ public class DropboxHelper {
 
   /**
    * Initializes the dropbox contexts.
+   *
    * @param accessToken The access token to use.
    */
   private void initDropBox(String accessToken) {
-    if (mDbxClient == null) {
-      DbxRequestConfig.Builder b = DbxRequestConfig.newBuilder(getClass().getPackage().getName());
+    final Package p = getClass().getPackage();
+    if (mDbxClient == null && p != null) {
+      DbxRequestConfig.Builder b = DbxRequestConfig.newBuilder(p.getName());
       DbxRequestConfig requestConfig = b.build();
       mDbxClient = new DbxClientV2(requestConfig, accessToken);
     }
@@ -97,6 +104,7 @@ public class DropboxHelper {
 
   /**
    * Returns the instance of the dropbox client.
+   *
    * @return DbxClientV2
    */
   DbxClientV2 getClient() {
