@@ -4,20 +4,21 @@ package fr.ralala.worktime.factories;
 import java.util.Collections;
 import java.util.List;
 
-import fr.ralala.worktime.models.PublicHolidayEntry;
 import fr.ralala.worktime.models.DayType;
+import fr.ralala.worktime.models.PublicHolidayEntry;
 import fr.ralala.worktime.models.WorkTimeDay;
 import fr.ralala.worktime.sql.SqlConstants;
 import fr.ralala.worktime.sql.SqlFactory;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Public holidays factory functions
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class PublicHolidaysFactory {
   private SqlFactory mSql = null;
@@ -25,6 +26,7 @@ public class PublicHolidaysFactory {
 
   /**
    * Sets the reference to the SqlFactory.
+   *
    * @param sql The SQLite factory.
    */
   public void setSqlFactory(final SqlFactory sql) {
@@ -33,27 +35,31 @@ public class PublicHolidaysFactory {
 
   /**
    * Returns the list of public holidays.
+   *
    * @return List<PublicHolidayEntry>
    */
   public List<PublicHolidayEntry> list() {
-    if(mSql == null)
+    if (mSql == null)
       return Collections.emptyList();
     return mSql.getPublicHolidays(-1, -1, -1);
   }
+
   /**
    * Returns the list of public holidays.
-   * @param year The public holiday year.
+   *
+   * @param year  The public holiday year.
    * @param month The public holiday month.
    * @return List<PublicHolidayEntry>
    */
   public List<PublicHolidayEntry> list(int year, int month) {
-    if(mSql == null)
+    if (mSql == null)
       return Collections.emptyList();
     return mSql.getPublicHolidays(year, month, -1);
   }
 
   /**
    * Returns the public holiday by name and date.
+   *
    * @param name The public holiday name.
    * @param date The public holiday date.
    * @return PublicHolidayEntry
@@ -64,13 +70,14 @@ public class PublicHolidaysFactory {
 
   /**
    * Tests id the current date is a public holiday.
-   * @param refList The reference List.
+   *
+   * @param refList     The reference List.
    * @param currentDate The current date.
    * @return boolean
    */
   public boolean isPublicHolidays(List<PublicHolidayEntry> refList, WorkTimeDay currentDate) {
-    for(PublicHolidayEntry de : refList) {
-      if((de.getTypeMorning() == DayType.PUBLIC_HOLIDAY && de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY) && de.matchSimpleDate(currentDate))
+    for (PublicHolidayEntry de : refList) {
+      if ((de.getTypeMorning() == DayType.PUBLIC_HOLIDAY && de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY) && de.matchSimpleDate(currentDate))
         return true;
     }
     return false;
@@ -78,17 +85,18 @@ public class PublicHolidaysFactory {
 
   /**
    * Tests the validity of the input entry.
+   *
    * @param phe The netry to test.
    * @return boolean
    */
   public boolean testValidity(final PublicHolidayEntry phe) {
-    WorkTimeDay de_day = phe.getDay();
-    List<PublicHolidayEntry> publicHolidays = mSql.getPublicHolidays(-1, de_day.getMonth(), de_day.getDay());
-    for(PublicHolidayEntry d : publicHolidays) {
-      WorkTimeDay d_day = d.getDay();
-      if(d.isRecurrence() || phe.isRecurrence()) {
+    WorkTimeDay deDay = phe.getDay();
+    List<PublicHolidayEntry> publicHolidays = mSql.getPublicHolidays(-1, deDay.getMonth(), deDay.getDay());
+    for (PublicHolidayEntry d : publicHolidays) {
+      WorkTimeDay dDay = d.getDay();
+      if (d.isRecurrence() || phe.isRecurrence()) {
         return false;
-      } else if(d_day.getYear() == de_day.getYear())
+      } else if (dDay.getYear() == deDay.getYear())
         return false;
     }
     return true;
@@ -96,6 +104,7 @@ public class PublicHolidaysFactory {
 
   /**
    * Removes an entry.
+   *
    * @param phe The entry to delete.
    */
   public void remove(final PublicHolidayEntry phe) {
@@ -104,6 +113,7 @@ public class PublicHolidaysFactory {
 
   /**
    * Adds a new entry.
+   *
    * @param phe The entry to add.
    */
   public void add(final PublicHolidayEntry phe) {
