@@ -4,61 +4,72 @@ package fr.ralala.worktime.models;
 import android.content.Context;
 
 import java.util.Calendar;
+import java.util.Objects;
 
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
 import fr.ralala.worktime.sql.SqlConstants;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Representation of a day entry
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class DayEntry {
+  private static final String TIME = "00:00";
   public static final int INVALID_WEEK = -1;
-  private WorkTimeDay mDay = null;
-  private WorkTimeDay mStartMorning = null;
-  private WorkTimeDay mEndMorning = null;
-  private WorkTimeDay mStartAfternoon = null;
-  private WorkTimeDay mEndAfternoon = null;
-  private WorkTimeDay mAdditionalBreak = null;
-  private WorkTimeDay mRecoveryTime = null;
-  private DayType mTypeMorning = DayType.ERROR;
-  private DayType mTypeAfternoon = DayType.ERROR;
+  private final WorkTimeDay mDay;
+  private final WorkTimeDay mStartMorning;
+  private final WorkTimeDay mEndMorning;
+  private final WorkTimeDay mStartAfternoon;
+  private final WorkTimeDay mEndAfternoon;
+  private final WorkTimeDay mAdditionalBreak;
+  private final WorkTimeDay mRecoveryTime;
+  private DayType mTypeMorning;
+  private DayType mTypeAfternoon;
   private double mAmountByHour = 0.0;
-  private WorkTimeDay mLegalWorktime = null;
-  private MainApplication mApp = null;
+  private WorkTimeDay mLegalWorktime;
+  private final MainApplication mApp;
   private long mId = SqlConstants.INVALID_ID;
   private int mWeekNumber = INVALID_WEEK;
 
+  @Override
+  public int hashCode() {
+    return Objects.hash(this);
+  }
+
   /**
    * Tests if an object is equal to this instance.
+   *
    * @param o The input object.
    * @return boolean
    */
+  @Override
   public boolean equals(Object o) {
-    if(o == null || !DayEntry.class.isInstance(o))
+    if (!(o instanceof DayEntry))
       return false;
     if (this == o)
       return true;
-    DayEntry de = (DayEntry)o;
+    DayEntry de = (DayEntry) o;
     return (mDay.match(de.mDay) && mStartMorning.match(de.mStartMorning)
-      && mEndMorning.match(de.mEndMorning)&& mAdditionalBreak.match(de.mAdditionalBreak)&& mRecoveryTime.match(de.mRecoveryTime) && mStartAfternoon.match(de.mStartAfternoon)
-      && mEndAfternoon.match(de.mEndAfternoon)&& mLegalWorktime.match(de.mLegalWorktime) && mTypeMorning == de.mTypeMorning
+      && mEndMorning.match(de.mEndMorning) && mAdditionalBreak.match(de.mAdditionalBreak) && mRecoveryTime.match(de.mRecoveryTime) && mStartAfternoon.match(de.mStartAfternoon)
+      && mEndAfternoon.match(de.mEndAfternoon) && mLegalWorktime.match(de.mLegalWorktime) && mTypeMorning == de.mTypeMorning
       && mTypeAfternoon == de.mTypeAfternoon && mAmountByHour == de.mAmountByHour);
   }
 
   /**
    * Constructs a new day entry.
-   * @param day Associated day.
-   * @param typeMorning Type of morning.
+   *
+   * @param day           Associated day.
+   * @param typeMorning   Type of morning.
    * @param typeAfternoon Type of afternoon.
    */
-  public DayEntry(final WorkTimeDay day, final DayType typeMorning, final DayType typeAfternoon) {
+  public DayEntry(Context ctx, final WorkTimeDay day, final DayType typeMorning, final DayType typeAfternoon) {
     mDay = new WorkTimeDay();
     mStartMorning = new WorkTimeDay();
     mEndMorning = new WorkTimeDay();
@@ -68,23 +79,25 @@ public class DayEntry {
     mEndAfternoon = new WorkTimeDay();
     mTypeMorning = typeMorning;
     mTypeAfternoon = typeAfternoon;
-    mApp = MainApplication.getInstance();
+    mApp = (MainApplication) ctx.getApplicationContext();
     mLegalWorktime = mApp.getLegalWorkTimeByDay();
     mDay.copy(day);
   }
 
   /**
    * Constructs a new day entry.
-   * @param day Associated day.
-   * @param typeMorning Type of morning.
+   *
+   * @param day           Associated day.
+   * @param typeMorning   Type of morning.
    * @param typeAfternoon Type of afternoon.
    */
-  public DayEntry(final Calendar day, final DayType typeMorning, final DayType typeAfternoon) {
-    this(new WorkTimeDay().fromCalendar(day), typeMorning, typeAfternoon);
+  public DayEntry(Context ctx, final Calendar day, final DayType typeMorning, final DayType typeAfternoon) {
+    this(ctx, new WorkTimeDay().fromCalendar(day), typeMorning, typeAfternoon);
   }
 
   /**
    * Sets the week number (if != INVALID_WEEK the item is invalidated and not clickable).
+   *
    * @param weekNumber The week number.
    */
   public void setWeekNumber(int weekNumber) {
@@ -93,6 +106,7 @@ public class DayEntry {
 
   /**
    * Returns the week number (if != INVALID_WEEK the item is invalidated and not clickable).
+   *
    * @return int
    */
   public int getWeekNumber() {
@@ -101,6 +115,7 @@ public class DayEntry {
 
   /**
    * Sets the DB id.
+   *
    * @param id The new ID.
    */
   public void setID(long id) {
@@ -109,6 +124,7 @@ public class DayEntry {
 
   /**
    * Returns the DB id.
+   *
    * @return long
    */
   public long getID() {
@@ -117,33 +133,36 @@ public class DayEntry {
 
   /**
    * Returns the work time pay.
+   *
    * @return double
    */
   public double getWorkTimePay(double defAmount) {
     double amount = getAmountByHour();
-    if(amount == 0 && defAmount != -1) amount = defAmount;
+    if (amount == 0 && defAmount != -1) amount = defAmount;
     return amount * Double.parseDouble(
-      String.valueOf(getWorkTime().getHours()) + "." + String.valueOf(getWorkTime().getMinutes()));
+      getWorkTime().getHours() + "." + getWorkTime().getMinutes());
   }
 
   /**
    * Returns the pause time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getPause() {
-    if(mTypeMorning == DayType.RECOVERY || mTypeAfternoon == DayType.RECOVERY)
+    if (mTypeMorning == DayType.RECOVERY || mTypeAfternoon == DayType.RECOVERY)
       return mAdditionalBreak;
-    if(mStartAfternoon.timeString().equals("00:00"))
+    if (mStartAfternoon.timeString().equals(TIME))
       return mStartAfternoon;
-    WorkTimeDay wp = mStartAfternoon.clone();
-    wp.delTime(mEndMorning.clone());
-    if(!mAdditionalBreak.timeString().equals("00:00"))
+    WorkTimeDay wp = mStartAfternoon.copy();
+    wp.delTime(mEndMorning.copy());
+    if (!mAdditionalBreak.timeString().equals(TIME))
       wp.addTime(mAdditionalBreak);
     return wp;
   }
 
   /**
    * Returns the work time pay.
+   *
    * @return double
    */
   public double getWorkTimePay() {
@@ -152,14 +171,16 @@ public class DayEntry {
 
   /**
    * Returns the over time (in milliseconds).
+   *
    * @return long
    */
   public long getOverTimeMs() {
-    return getWorkTime().getTimeMs() - getRealLegalWorktimeMS().getTimeMs();
+    return getWorkTime().getTimeMs() - getRealLegalWorkTimeMS().getTimeMs();
   }
 
   /**
    * Returns the over time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getOverTime(long diffInMs) {
@@ -168,6 +189,7 @@ public class DayEntry {
 
   /**
    * Returns the over time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getOverTime() {
@@ -176,24 +198,26 @@ public class DayEntry {
 
   /**
    * Returns the work time value.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getWorkTime() {
-    WorkTimeDay wm = (getTypeMorning() != DayType.RECOVERY && isValidMorningType()) ? mEndMorning.clone() : new WorkTimeDay();
-    if(!wm.timeString().equals("00:00") && isValidMorningType() && getTypeMorning() != DayType.RECOVERY)
+    WorkTimeDay wm = (getTypeMorning() != DayType.RECOVERY && isValidMorningType()) ? mEndMorning.copy() : new WorkTimeDay();
+    if (!wm.timeString().equals(TIME) && isValidMorningType() && getTypeMorning() != DayType.RECOVERY)
       wm.delTime(mStartMorning);
-    WorkTimeDay wa = (getTypeAfternoon() != DayType.RECOVERY && isValidAfternoonType()) ? mEndAfternoon.clone() : new WorkTimeDay();
-    if(!wa.timeString().equals("00:00") && isValidAfternoonType() && getTypeAfternoon() != DayType.RECOVERY)
+    WorkTimeDay wa = (getTypeAfternoon() != DayType.RECOVERY && isValidAfternoonType()) ? mEndAfternoon.copy() : new WorkTimeDay();
+    if (!wa.timeString().equals(TIME) && isValidAfternoonType() && getTypeAfternoon() != DayType.RECOVERY)
       wa.delTime(mStartAfternoon);
-    WorkTimeDay wt = wm.clone();
+    WorkTimeDay wt = wm.copy();
     wt.addTime(wa);
-    if(!wt.timeString().equals("00:00") && !mAdditionalBreak.timeString().equals("00:00"))
-      wt.delTime(mAdditionalBreak.clone());
+    if (!wt.timeString().equals(TIME) && !mAdditionalBreak.timeString().equals(TIME))
+      wt.delTime(mAdditionalBreak.copy());
     return wt;
   }
 
   /**
    * Copies another object to this object.
+   *
    * @param de The object to copy.
    */
   public void copy(DayEntry de) {
@@ -213,20 +237,22 @@ public class DayEntry {
 
   /**
    * Tests if the current instance matches with the current entry.
-   * @param de The entry to test.
+   *
+   * @param de       The entry to test.
    * @param testName Test name.
    * @return boolean
    */
   public boolean match(DayEntry de, boolean testName) {
     return !(testName) && mDay.match(de.mDay) &&
-        mStartMorning.match(de.mStartMorning) && mEndMorning.match(de.mEndMorning) &&
-        mAdditionalBreak.match(de.mAdditionalBreak) && mRecoveryTime.match(de.mRecoveryTime) && mStartAfternoon.match(de.mStartAfternoon) &&
-        mEndAfternoon.match(de.mEndAfternoon) && mTypeMorning == de.mTypeMorning && mTypeAfternoon ==
-        de.mTypeAfternoon && mAmountByHour == de.mAmountByHour && mLegalWorktime.match(de.mLegalWorktime);
+      mStartMorning.match(de.mStartMorning) && mEndMorning.match(de.mEndMorning) &&
+      mAdditionalBreak.match(de.mAdditionalBreak) && mRecoveryTime.match(de.mRecoveryTime) && mStartAfternoon.match(de.mStartAfternoon) &&
+      mEndAfternoon.match(de.mEndAfternoon) && mTypeMorning == de.mTypeMorning && mTypeAfternoon ==
+      de.mTypeAfternoon && mAmountByHour == de.mAmountByHour && mLegalWorktime.match(de.mLegalWorktime);
   }
 
   /**
    * Tests if the date matches with the current entry.
+   *
    * @param current The entry to test.
    * @return boolean
    */
@@ -236,33 +262,45 @@ public class DayEntry {
 
   /**
    * Returns the day string in 2 letters.
-   * @param c The Android context.
+   *
+   * @param c         The Android context.
    * @param dayOfWeek The day of week (See Calendar.MONDAY, etc...)
    * @return String
    */
   public static String getDayString2lt(final Context c, final int dayOfWeek) {
     switch (dayOfWeek) {
-      case Calendar.MONDAY: return c.getString(R.string.day_monday_2lt);
-      case Calendar.TUESDAY: return c.getString(R.string.day_tuesday_2lt);
-      case Calendar.WEDNESDAY: return c.getString(R.string.day_wednesday_2lt);
-      case Calendar.THURSDAY: return c.getString(R.string.day_thursday_2lt);
-      case Calendar.FRIDAY: return c.getString(R.string.day_friday_2lt);
-      case Calendar.SATURDAY: return c.getString(R.string.day_saturday_2lt);
-      case Calendar.SUNDAY: return c.getString(R.string.day_sunday_2lt);
+      case Calendar.MONDAY:
+        return c.getString(R.string.day_monday_2lt);
+      case Calendar.TUESDAY:
+        return c.getString(R.string.day_tuesday_2lt);
+      case Calendar.WEDNESDAY:
+        return c.getString(R.string.day_wednesday_2lt);
+      case Calendar.THURSDAY:
+        return c.getString(R.string.day_thursday_2lt);
+      case Calendar.FRIDAY:
+        return c.getString(R.string.day_friday_2lt);
+      case Calendar.SATURDAY:
+        return c.getString(R.string.day_saturday_2lt);
+      case Calendar.SUNDAY:
+        return c.getString(R.string.day_sunday_2lt);
+      default:
+        break;
     }
     return c.getString(R.string.error);
   }
 
   /**
    * Parses the input time.
+   *
    * @param time The input time.
    * @return WorkTimeDay
    */
   private WorkTimeDay parseTime(String time) {
-    if(time == null)
-      time = "00:00";
+    String t = time;
+    if (t == null)
+      t = TIME;
     WorkTimeDay wtd = new WorkTimeDay();
-    String split [] = time.split(":");
+    String[] split = t.split(":");
     wtd.setHours(Integer.parseInt(split[0]));
     wtd.setMinutes(Integer.parseInt(split[1]));
     return wtd;
@@ -270,12 +308,13 @@ public class DayEntry {
 
   /**
    * Parses the input date.
+   *
    * @param date The input date.
    * @return WorkTimeDay
    */
   private WorkTimeDay parseDate(String date) {
     WorkTimeDay wtd = new WorkTimeDay();
-    String split [] = date.split("/");
+    String[] split = date.split("/");
     wtd.setDay(Integer.parseInt(split[0]));
     wtd.setMonth(Integer.parseInt(split[1]));
     wtd.setYear(Integer.parseInt(split[2]));
@@ -284,6 +323,7 @@ public class DayEntry {
 
   /**
    * Returns the start morning time.
+   *
    * @param start The new value.
    */
   public void setStartMorning(String start) {
@@ -292,6 +332,7 @@ public class DayEntry {
 
   /**
    * Returns the start afternoon time.
+   *
    * @param start The new value.
    */
   public void setStartAfternoon(String start) {
@@ -300,6 +341,7 @@ public class DayEntry {
 
   /**
    * Sets the end morning time.
+   *
    * @param end The new value.
    */
   public void setEndMorning(String end) {
@@ -308,6 +350,7 @@ public class DayEntry {
 
   /**
    * Sets the end afternoon time.
+   *
    * @param end The new value.
    */
   public void setEndAfternoon(String end) {
@@ -316,6 +359,7 @@ public class DayEntry {
 
   /**
    * Sets the end afternoon time.
+   *
    * @param end The new value.
    */
   public void setEndAfternoon(WorkTimeDay end) {
@@ -324,6 +368,7 @@ public class DayEntry {
 
   /**
    * Sets the end morning time.
+   *
    * @param end The new value.
    */
   public void setEndMorning(WorkTimeDay end) {
@@ -333,6 +378,7 @@ public class DayEntry {
 
   /**
    * Sets the additional break time.
+   *
    * @param sbreak The new time in String.
    */
   public void setAdditionalBreak(String sbreak) {
@@ -341,6 +387,7 @@ public class DayEntry {
 
   /**
    * Returns the additional break time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getAdditionalBreak() {
@@ -349,6 +396,7 @@ public class DayEntry {
 
   /**
    * Sets the recovery time.
+   *
    * @param sbreak The new time in String.
    */
   public void setRecoveryTime(String sbreak) {
@@ -357,6 +405,7 @@ public class DayEntry {
 
   /**
    * Returns the recovery time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getRecoveryTime() {
@@ -365,6 +414,7 @@ public class DayEntry {
 
   /**
    * Sets the day.
+   *
    * @param day The new day.
    */
   public void setDay(String day) {
@@ -373,6 +423,7 @@ public class DayEntry {
 
   /**
    * Returns the day.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getDay() {
@@ -381,13 +432,16 @@ public class DayEntry {
 
   /**
    * Returns the start morning time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getStartMorning() {
     return mStartMorning;
   }
+
   /**
    * Returns the start afternoon time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getStartAfternoon() {
@@ -397,6 +451,7 @@ public class DayEntry {
 
   /**
    * Tests if the morning type is valid.
+   *
    * @return boolean
    */
   public boolean isValidMorningType() {
@@ -405,6 +460,7 @@ public class DayEntry {
 
   /**
    * Tests if the afternoon type is valid.
+   *
    * @return boolean
    */
   public boolean isValidAfternoonType() {
@@ -413,6 +469,7 @@ public class DayEntry {
 
   /**
    * Returns the type of the morning time.
+   *
    * @return DayType
    */
   public DayType getTypeMorning() {
@@ -421,6 +478,7 @@ public class DayEntry {
 
   /**
    * Sets the type of the morning time.
+   *
    * @param type The new type.
    */
   public void setTypeMorning(DayType type) {
@@ -429,6 +487,7 @@ public class DayEntry {
 
   /**
    * Returns the type of the afternoon time.
+   *
    * @return DayType
    */
   public DayType getTypeAfternoon() {
@@ -437,6 +496,7 @@ public class DayEntry {
 
   /**
    * Sets the type of the afternoon time.
+   *
    * @param type The new type.
    */
   public void setTypeAfternoon(DayType type) {
@@ -445,6 +505,7 @@ public class DayEntry {
 
   /**
    * Returns the end morning time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getEndMorning() {
@@ -453,6 +514,7 @@ public class DayEntry {
 
   /**
    * Returns the end afternoon time.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getEndAfternoon() {
@@ -461,6 +523,7 @@ public class DayEntry {
 
   /**
    * Returns the amount by hour.
+   *
    * @return double
    */
   public double getAmountByHour() {
@@ -469,6 +532,7 @@ public class DayEntry {
 
   /**
    * Sets the amount by hour.
+   *
    * @param amountByHour The new value.
    */
   public void setAmountByHour(double amountByHour) {
@@ -478,10 +542,11 @@ public class DayEntry {
 
   /**
    * Returns the legal work time value.
+   *
    * @return WorkTimeDay
    */
-  public WorkTimeDay getLegalWorktime() {
-    if(mLegalWorktime.timeString().equals("00:00"))
+  public WorkTimeDay getLegalWorkTime() {
+    if (mLegalWorktime.timeString().equals(TIME))
       mLegalWorktime = mApp.getLegalWorkTimeByDay();
     return mLegalWorktime;
   }
@@ -489,23 +554,25 @@ public class DayEntry {
 
   /**
    * Returns the legal work time value adjusted with the recovery time.
+   *
    * @return WorkTimeDay
    */
-  private WorkTimeDay getRealLegalWorktimeMS() {
-    WorkTimeDay wmLegal = getLegalWorktime();
+  private WorkTimeDay getRealLegalWorkTimeMS() {
+    WorkTimeDay wmLegal = getLegalWorkTime();
     WorkTimeDay wm = new WorkTimeDay();
-    if(!wmLegal.timeString().equals("00:00"))
+    if (!wmLegal.timeString().equals(TIME))
       wm.addTime(wmLegal);
-    if(!wm.timeString().equals("00:00") && !mRecoveryTime.timeString().equals("00:00"))
-      wm.delTime(mRecoveryTime.clone());
+    if (!wm.timeString().equals(TIME) && !mRecoveryTime.timeString().equals(TIME))
+      wm.delTime(mRecoveryTime.copy());
     return wm;
   }
 
   /**
    * Sets the legal work time from a string value.
-   * @param legalWorktime The new value.
+   *
+   * @param legalWorkTime The new value.
    */
-  public void setLegalWorktime(String legalWorktime) {
-    mLegalWorktime.copy(parseTime(legalWorktime));
+  public void setLegalWorkTime(String legalWorkTime) {
+    mLegalWorktime.copy(parseTime(legalWorkTime));
   }
 }
