@@ -1,31 +1,37 @@
 package fr.ralala.worktime.models;
 
+import android.content.Context;
+
 import fr.ralala.worktime.sql.SqlConstants;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Representation of a public holiday entry
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class PublicHolidayEntry {
   private String mName = "";
-  private WorkTimeDay mDay;
-  private DayType mTypeMorning;
-  private DayType mTypeAfternoon;
+  private final WorkTimeDay mDay;
+  private final DayType mTypeMorning;
+  private final DayType mTypeAfternoon;
   private boolean mRecurrence = false;
   private long mId = SqlConstants.INVALID_ID;
+  private final Context mCtx;
 
   /**
    * Constructs a new day entry.
-   * @param day Associated day.
-   * @param typeMorning Type of morning.
+   *
+   * @param day           Associated day.
+   * @param typeMorning   Type of morning.
    * @param typeAfternoon Type of afternoon.
    */
-  public PublicHolidayEntry(final WorkTimeDay day, final DayType typeMorning, final DayType typeAfternoon) {
+  public PublicHolidayEntry(Context ctx, final WorkTimeDay day, final DayType typeMorning, final DayType typeAfternoon) {
+    mCtx = ctx;
     mDay = new WorkTimeDay();
     mTypeMorning = typeMorning;
     mTypeAfternoon = typeAfternoon;
@@ -34,14 +40,16 @@ public class PublicHolidayEntry {
 
   /**
    * Converts the PublicHolidayEntry to DayEntry.
+   *
    * @return DayEntry
    */
   public DayEntry toDayEntry() {
-    return new DayEntry(mDay, mTypeMorning, mTypeAfternoon);
+    return new DayEntry(mCtx, mDay, mTypeMorning, mTypeAfternoon);
   }
 
   /**
    * Sets the DB id.
+   *
    * @param id The new ID.
    */
   public void setID(long id) {
@@ -50,6 +58,7 @@ public class PublicHolidayEntry {
 
   /**
    * Returns the DB id.
+   *
    * @return long
    */
   public long getID() {
@@ -58,19 +67,21 @@ public class PublicHolidayEntry {
 
   /**
    * Tests if the date matches with the current entry.
+   *
    * @param current The entry to test.
    * @return boolean
    */
   public boolean matchSimpleDate(WorkTimeDay current) {
     boolean ret = (current.getMonth() == mDay.getMonth() && current.getDay() == mDay.getDay());
     /* simple holidays can change between each years */
-    if(!isRecurrence() && ret)
+    if (!isRecurrence() && ret)
       return current.getYear() == mDay.getYear();
     return ret;
   }
 
   /**
    * Returns the day.
+   *
    * @return WorkTimeDay
    */
   public WorkTimeDay getDay() {
@@ -79,6 +90,7 @@ public class PublicHolidayEntry {
 
   /**
    * Returns the type of the morning time.
+   *
    * @return DayType
    */
   public DayType getTypeMorning() {
@@ -87,6 +99,7 @@ public class PublicHolidayEntry {
 
   /**
    * Returns the type of the afternoon time.
+   *
    * @return DayType
    */
   public DayType getTypeAfternoon() {
@@ -95,6 +108,7 @@ public class PublicHolidayEntry {
 
   /**
    * Returns the name (if used with profile or public holidays.
+   *
    * @return String
    */
   public String getName() {
@@ -103,6 +117,7 @@ public class PublicHolidayEntry {
 
   /**
    * Sets the name (if used with profile or public holidays)
+   *
    * @param name The new value.
    */
   public void setName(String name) {
@@ -111,6 +126,7 @@ public class PublicHolidayEntry {
 
   /**
    * Tests if the recurrence state is enabled.
+   *
    * @return boolean
    */
   public boolean isRecurrence() {
@@ -119,6 +135,7 @@ public class PublicHolidayEntry {
 
   /**
    * Sets the recurrence state.
+   *
    * @param recurrence The new state.
    */
   public void setRecurrence(boolean recurrence) {
