@@ -4,49 +4,55 @@ package fr.ralala.worktime.ui.activities;
 import android.Manifest;
 import android.content.ComponentCallbacks2;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.os.Process;
-import android.os.Vibrator;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.Toolbar;
-import android.view.Gravity;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.Window;
 
+import androidx.activity.OnBackPressedCallback;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.google.android.material.navigation.NavigationView;
+
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
+import fr.ralala.worktime.launchers.LauncherDayActivity;
+import fr.ralala.worktime.launchers.LauncherPublicHolidayActivity;
+import fr.ralala.worktime.services.AutoExportService;
 import fr.ralala.worktime.ui.activities.settings.SettingsActivity;
 import fr.ralala.worktime.ui.changelog.ChangeLog;
 import fr.ralala.worktime.ui.changelog.ChangeLogIds;
 import fr.ralala.worktime.ui.fragments.AppFragmentsFactory;
 import fr.ralala.worktime.ui.fragments.WorkTimeFragment;
-import fr.ralala.worktime.services.DropboxAutoExportService;
-import fr.ralala.worktime.services.QuickAccessService;
-import fr.ralala.worktime.utils.AndroidHelper;
 import fr.ralala.worktime.ui.utils.SwipeDetector;
 import fr.ralala.worktime.ui.utils.UIHelper;
+import fr.ralala.worktime.utils.AndroidHelper;
 import fr.ralala.worktime.utils.DefaultExceptionHandler;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Main activity, manage the navigation view
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
-public class MainActivity extends RuntimePermissionsActivity implements NavigationView.OnNavigationItemSelectedListener, ComponentCallbacks2 {
-  private static final int PERMISSIONS_REQUEST = 30;
+public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener, ComponentCallbacks2 {
   private static final int BACK_TIME_DELAY = 2000;
   private static long mLastBackPressed = -1;
   private boolean mViewIsAtHome = false;
@@ -54,14 +60,16 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
   private NavigationView mNavigationView = null;
   private SwipeDetector mSwipeDetector = null;
   private DrawerLayout mDrawer = null;
-  private AppFragmentsFactory mFragments = null;
+  private AppFragmentsFactory mFragmentsFactory = null;
   private AlertDialog mProgress;
   private boolean mLockedByProgress = false;
   private boolean mResumeFromActivity = false;
-  private Vibrator mVibrator;
+  private LauncherDayActivity mLauncherDayActivity;
+  private LauncherPublicHolidayActivity mLauncherPublicHolidayActivity;
 
   /**
    * Called when the activity is created.
+   *
    * @param savedInstanceState The saved instance state.
    */
   @Override
@@ -70,7 +78,6 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
     DefaultExceptionHandler.install(this);
     setContentView(R.layout.activity_main);
 
-    mVibrator = ((Vibrator) getSystemService(VIBRATOR_SERVICE));
     Intent intent = getIntent();
     if (intent.getExtras() != null && intent.hasExtra(AndroidHelper.EXTRA_RESTART)) {
       String msg = intent.getExtras().getString(AndroidHelper.EXTRA_RESTART);
@@ -87,23 +94,27 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
     toggle.syncState();
 
     mNavigationView = findViewById(R.id.nav_view);
-    if(mNavigationView != null) {
+    if (mNavigationView != null) {
       mNavigationView.setNavigationItemSelectedListener(this);
       mNavigationView.getMenu().getItem(0).setChecked(true);
     }
-    mApp = MainApplication.getInstance();
-    mFragments = new AppFragmentsFactory(mApp, mNavigationView);
+    mApp = (MainApplication) getApplication();
+    mFragmentsFactory = new AppFragmentsFactory(mApp, mNavigationView);
 
-    String[] perms = new String[]{
+    /* permissions */
+    String[] permissions = new String[]{
       Manifest.permission.READ_EXTERNAL_STORAGE,
       Manifest.permission.WRITE_EXTERNAL_STORAGE,
       Manifest.permission.INTERNET,
-      Manifest.permission.VIBRATE,
       Manifest.permission.ACCESS_FINE_LOCATION,
       Manifest.permission.ACCESS_COARSE_LOCATION,
-      Manifest.permission.VIBRATE
+      Manifest.permission.VIBRATE,
     };
-    super.requestAppPermissions(perms, R.string.permissions_read_ext_storage , PERMISSIONS_REQUEST);
+    boolean perms = true;
+    for (String perm : permissions)
+      perms &= ContextCompat.checkSelfPermission(this, perm) == PackageManager.PERMISSION_GRANTED;
+    if (!perms)
+      ActivityCompat.requestPermissions(this, permissions, 1);
     ChangeLog changeLog = new ChangeLog(
       new ChangeLogIds(
         R.raw.changelog,
@@ -112,12 +123,54 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
         R.string.changelog_title,
         R.string.changelog_full_title,
         R.string.changelog_show_full), this);
-    if(changeLog.firstRun())
+    if (changeLog.firstRun())
       changeLog.getLogDialog().show();
+
+    mLauncherDayActivity = new LauncherDayActivity(this);
+    mLauncherPublicHolidayActivity = new LauncherPublicHolidayActivity(this);
+    getOnBackPressedDispatcher().addCallback(this, new OnBackPressedCallback(true) {
+      @Override
+      public void handleOnBackPressed() {
+        back();
+      }
+    });
+  }
+
+  private void back() {
+    if (!mFragmentsFactory.consumeBackPressed()) {
+      if (mViewIsAtHome)
+        mDrawer.openDrawer(GravityCompat.START);
+      else { //if the current view is not the News fragment
+        displayView(mFragmentsFactory.getDefaultHomeId()); //display the home fragment
+        mNavigationView.getMenu().getItem(mFragmentsFactory.getDefaultHomeIndex()).setChecked(true); /* select home title */
+        return;
+      }
+      if (mLastBackPressed + BACK_TIME_DELAY > System.currentTimeMillis()) {
+        cleanup();
+        finish();
+        return;
+      } else {
+        UIHelper.toast(this, R.string.on_double_back_exit_text);
+      }
+      updateLastBackPressed();
+    }
+  }
+
+  private static void updateLastBackPressed() {
+    mLastBackPressed = System.currentTimeMillis();
+  }
+
+  public LauncherPublicHolidayActivity getLauncherPublicHolidayActivity() {
+    return mLauncherPublicHolidayActivity;
+  }
+
+  public LauncherDayActivity getLauncherDayActivity() {
+    return mLauncherDayActivity;
   }
 
   /**
    * Called to create the option menu.
+   *
    * @param menu The main menu.
    * @return boolean
    */
@@ -128,9 +181,9 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
   }
 
 
-
   /**
    * Called when the user select an option menu item.
+   *
    * @param item The selected item.
    * @return boolean
    */
@@ -138,9 +191,7 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
   public boolean onOptionsItemSelected(final MenuItem item) {
     final int id = item.getItemId();
     if (id == R.id.action_last_export) {
-      if(mApp.getLastExportType().equals(MainApplication.PREFS_VAL_LAST_EXPORT_DEVICE)) {
-        AndroidHelper.exportDevice(this);
-      } else if(mApp.getLastExportType().equals(MainApplication.PREFS_VAL_LAST_EXPORT_DROPBOX)) {
+      if (mApp.getLastExportType().equals(MainApplication.PREFS_VAL_LAST_EXPORT_DROPBOX)) {
         AndroidHelper.exportDropbox(mApp, this);
       }
       return true;
@@ -149,33 +200,11 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
   }
 
   /**
-   * Called when the file chooser is disposed with a result.
-   * @param requestCode The request code.
-   * @param resultCode The result code.
-   * @param data The Intent data.
-   */
-  @Override
-  protected void onActivityResult(final int requestCode, final int resultCode, final Intent data) {
-    // Check which request we're responding to
-    if(!AndroidHelper.exportDeviceActivityResult(this, requestCode, resultCode, data))
-      // the event is not consumed by the activity, it must be propagated to fragments
-      super.onActivityResult(requestCode, resultCode, data);
-  }
-
-  /**
-   * Returns the instance of the vibrator service.
-   * @return Vibrator
-   */
-  public Vibrator getVibrator() {
-    return mVibrator;
-  }
-
-  /**
    * Show the progress dialog.
    */
   public void progressShow(boolean locked) {
     mLockedByProgress = locked;
-    if(!mProgress.isShowing()) {
+    if (!mProgress.isShowing()) {
       mProgress.show();
       Window window = mProgress.getWindow();
       if (window != null) {
@@ -191,27 +220,18 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
    */
   public void progressDismiss() {
     mLockedByProgress = false;
-    if(mProgress.isShowing()) {
+    if (mProgress.isShowing()) {
       mProgress.dismiss();
     }
   }
 
   /**
    * Returns the instance of the SwipeDetector.
+   *
    * @return SwipeDetector
    */
   public SwipeDetector getSwipeDetector() {
     return mSwipeDetector;
-  }
-
-  /**
-   * Called when the permissions are granted.
-   * @param requestCode The request code.
-   */
-  @Override
-  public void onPermissionsGranted(final int requestCode) {
-    //AndroidHelper.toastLong(this, R.string.permissions_read_ext_storage_done);
-
   }
 
   /**
@@ -220,28 +240,26 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
   @Override
   public void onResume() {
     super.onResume();
-    mApp.incResumedCounter();
-    if(mFragments.isRequiresProgress() && !mResumeFromActivity && mApp.getResumedCounter() == 1)
+    MainApplication.incResumedCounter();
+    if (mFragmentsFactory.isRequiresProgress() && !mResumeFromActivity && MainApplication.getResumedCounter() == 1)
       progressShow(true);
     new Thread(() -> {
-     /* load SQL */
-      if(!mApp.openSql(this)) finish();
-      /*if(mApp.isExportAutoSave())
-        mApp.initOnLoadTables();*/
+      /* load SQL */
+      if (!mApp.openSql(this)) finish();
       runOnUiThread(() -> {
-        displayView(mFragments.getCurrentFragmentId());
-        if(mApp.getLastWidgetOpen() != 0L) {
+        displayView(mFragmentsFactory.getCurrentFragmentId());
+        if (mApp.getLastWidgetOpen() != 0L) {
           long elapsed = System.currentTimeMillis() - mApp.getLastWidgetOpen();
-          if(elapsed <= 500) {
+          if (elapsed <= 500) {
             mApp.setLastWidgetOpen(0L);
             progressDismiss();
             finish();
             return;
           }
         }
-        AndroidHelper.killServiceIfRunning(this, DropboxAutoExportService.class);
-        mFragments.onResume(this);
-        if(!mLockedByProgress && mProgress.isShowing())
+        AndroidHelper.killServiceIfRunning(this, AutoExportService.class);
+        mFragmentsFactory.onResume(this);
+        if (!mLockedByProgress && mProgress.isShowing())
           mProgress.dismiss();
       });
 
@@ -251,6 +269,7 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
 
   /**
    * Sets resume from activity state.
+   *
    * @param resumeFromActivity The new state.
    */
   public void setResumeFromActivity(boolean resumeFromActivity) {
@@ -270,49 +289,22 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
    * Cleanup the resources.
    */
   private void cleanup() {
-    if(mApp != null) {
-      if(mProgress.isShowing())
+    if (mApp != null) {
+      if (mProgress.isShowing())
         mProgress.dismiss();
-      if(mApp.isExportAutoSave()) {
-        startService(new Intent(this, DropboxAutoExportService.class));
+      if (mApp.isExportAutoSave()) {
+        startService(new Intent(this, AutoExportService.class));
       }
       mApp.getSql().close();
-      mApp.getQuickAccessNotification().remove(this);
-    }
-    AndroidHelper.killServiceIfRunning(this, QuickAccessService.class);
-  }
-
-  /**
-   * Called to handle the click on the back button.
-   */
-  @Override
-  public void onBackPressed() {
-    if(!mFragments.consumeBackPressed()) {
-      if (mViewIsAtHome)
-        mDrawer.openDrawer(Gravity.START);
-      else { //if the current view is not the News fragment
-        displayView(mFragments.getDefaultHomeId()); //display the home fragment
-        mNavigationView.getMenu().getItem(mFragments.getDefaultHomeIndex()).setChecked(true); /* select home title */
-        return;
-      }
-      if (mLastBackPressed + BACK_TIME_DELAY > System.currentTimeMillis()) {
-        cleanup();
-        super.onBackPressed();
-        //Process.killProcess(android.os.Process.myPid());
-        return;
-      } else {
-        UIHelper.toast(this, R.string.on_double_back_exit_text);
-      }
-      mLastBackPressed = System.currentTimeMillis();
     }
   }
 
   /**
    * Called when an item is selected in the navigation view.
+   *
    * @param item The selected item.
    * @return boolean
    */
-  @SuppressWarnings("StatementWithEmptyBody")
   @Override
   public boolean onNavigationItemSelected(@NonNull MenuItem item) {
     displayView(item.getItemId());
@@ -321,69 +313,60 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
 
   /**
    * Dispatch the touch event.
+   *
    * @param ev The event.
    * @return false if the event is not consumed.
    */
   @Override
-  public boolean dispatchTouchEvent(MotionEvent ev){
+  public boolean dispatchTouchEvent(MotionEvent ev) {
     boolean b = super.dispatchTouchEvent(ev);
-    return WorkTimeFragment.class.isInstance(mFragments.getCurrentFragment()) ? mSwipeDetector.onTouchEvent(ev) : b;
+    return mFragmentsFactory.getCurrentFragment() instanceof WorkTimeFragment ? mSwipeDetector.onTouchEvent(ev) : b;
   }
 
   /**
    * Displays a specific view.
+   *
    * @param viewId The view id to display.
    */
   public void displayView(int viewId) {
     String title = getString(R.string.app_title);
 
-    mViewIsAtHome = mFragments.getDefaultHomeId() == viewId;
-    switch (viewId) {
-      case R.id.nav_quickaccess:
-        mFragments.setCurrentToFragment(AppFragmentsFactory.IDX_QUICK_ACCESS);
-        title = getString(R.string.quickaccess);
-        break;
-      case R.id.nav_profile:
-        mFragments.setCurrentToFragment(AppFragmentsFactory.IDX_PROFILE);
-        title  = getString(R.string.profile);
-        break;
-      case R.id.nav_public_holidays:
-        mFragments.setCurrentToFragment(AppFragmentsFactory.IDX_PUBLIC_HOLIDAY);
-        title = getString(R.string.public_holidays);
-        break;
-      case R.id.nav_export:
-        mFragments.setCurrentToFragment(AppFragmentsFactory.IDX_EXPORT);
-        title = getString(R.string.export);
-        break;
-      case R.id.nav_statistics:
-        mFragments.setCurrentToFragment(AppFragmentsFactory.IDX_STATISTICS);
-        title = getString(R.string.statistics);
-        break;
-      case R.id.nav_worktime:
-        mFragments.setCurrentToFragment(AppFragmentsFactory.IDX_WORK_TIME);
-        title = getString(R.string.work_time);
-        break;
-      case R.id.nav_settings:
-        setResumeFromActivity(true);
-        mFragments.setCurrentToFragment(-1);
-        startActivity(new Intent(this, SettingsActivity.class));
-        if(mDrawer != null) mDrawer.closeDrawer(GravityCompat.START);
-        return;
-      case R.id.nav_exit:
-        cleanup();
-        super.onBackPressed();
-        Process.killProcess(android.os.Process.myPid());
-        return;
-      default:
-        mFragments.setCurrentToFragment(-1);
-        break;
+    mViewIsAtHome = mFragmentsFactory.getDefaultHomeId() == viewId;
 
+    if (viewId == R.id.nav_profile) {
+      mFragmentsFactory.setCurrentToFragment(AppFragmentsFactory.IDX_PROFILE);
+      title = getString(R.string.profile);
+    } else if (viewId == R.id.nav_public_holidays) {
+      mFragmentsFactory.setCurrentToFragment(AppFragmentsFactory.IDX_PUBLIC_HOLIDAY);
+      title = getString(R.string.public_holidays);
+    } else if (viewId == R.id.nav_export) {
+      mFragmentsFactory.setCurrentToFragment(AppFragmentsFactory.IDX_EXPORT);
+      title = getString(R.string.export);
+    } else if (viewId == R.id.nav_statistics) {
+      mFragmentsFactory.setCurrentToFragment(AppFragmentsFactory.IDX_STATISTICS);
+      title = getString(R.string.statistics);
+    } else if (viewId == R.id.nav_worktime) {
+      mFragmentsFactory.setCurrentToFragment(AppFragmentsFactory.IDX_WORK_TIME);
+      title = getString(R.string.work_time);
+    } else if (viewId == R.id.nav_settings) {
+      setResumeFromActivity(true);
+      mFragmentsFactory.setCurrentToFragment(-1);
+      startActivity(new Intent(this, SettingsActivity.class));
+      if (mDrawer != null) mDrawer.closeDrawer(GravityCompat.START);
+      return;
+    } else if (viewId == R.id.nav_exit) {
+      cleanup();
+      back();
+      Process.killProcess(android.os.Process.myPid());
+      return;
+    } else {
+      mFragmentsFactory.setCurrentToFragment(-1);
     }
 
-    if (mFragments.getCurrentFragment() != null) {
+    if (mFragmentsFactory.getCurrentFragment() != null) {
       FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
       ft.setCustomAnimations(R.anim.fadein, R.anim.fadeout);
-      ft.replace(R.id.content_frame, mFragments.getCurrentFragment());
+      ft.replace(R.id.content_frame, mFragmentsFactory.getCurrentFragment());
       ft.commit();
     }
 
@@ -392,7 +375,7 @@ public class MainActivity extends RuntimePermissionsActivity implements Navigati
       getSupportActionBar().setTitle(title);
     }
 
-    if(mDrawer != null) mDrawer.closeDrawer(GravityCompat.START);
+    if (mDrawer != null) mDrawer.closeDrawer(GravityCompat.START);
 
   }
 }
