@@ -3,7 +3,6 @@ package fr.ralala.worktime.ui.adapters;
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.Typeface;
-import android.support.annotation.NonNull;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -11,6 +10,8 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
 
 import java.util.Calendar;
 import java.util.List;
@@ -30,16 +31,16 @@ import fr.ralala.worktime.ui.utils.UIHelper;
  * </p>
  *
  * @author Keidan
- *         <p>
- *******************************************************************************
+ * <p>
+ * ******************************************************************************
  */
 public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
+  private static final String TIME_ZERO = "00:00";
+  private final Context mContext;
+  private final int mId;
+  private final List<DayEntry> mItems;
 
-  private Context mContext;
-  private int mId;
-  private List<DayEntry> mItems;
-
-  private class ViewHolder {
+  private static class ViewHolder {
     LinearLayout llDay = null;
     TextView tvWeek = null;
     TextView tvDay = null;
@@ -53,12 +54,13 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
 
   /**
    * Creates the array adapter.
-   * @param context The Android context.
+   *
+   * @param context            The Android context.
    * @param textViewResourceId The resource id of the container.
-   * @param objects The objects list.
+   * @param objects            The objects list.
    */
   public DaysEntriesArrayAdapter(final Context context, final int textViewResourceId,
-                                           final List<DayEntry> objects) {
+                                 final List<DayEntry> objects) {
     super(context, textViewResourceId, objects);
     mContext = context;
     mId = textViewResourceId;
@@ -67,6 +69,7 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
 
   /**
    * Returns an items at a specific position.
+   *
    * @param i The item index.
    * @return The item.
    */
@@ -77,14 +80,15 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
 
   /**
    * Returns the current view.
-   * @param position The view position.
+   *
+   * @param position    The view position.
    * @param convertView The view to convert.
-   * @param parent The parent.
+   * @param parent      The parent.
    * @return The new view.
    */
   @Override
   public @NonNull View getView(final int position, final View convertView,
-                      @NonNull final ViewGroup parent) {
+                               @NonNull final ViewGroup parent) {
     View v = convertView;
     ViewHolder holder;
     final DayEntry t = mItems.get(position);
@@ -104,16 +108,17 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
       holder.tvOverColors = holder.tvOver.getTextColors();
       v.setTag(holder);
     } else {
-        /* We recycle a View that already exists */
+      /* We recycle a View that already exists */
       holder = (ViewHolder) v.getTag();
     }
 
     WorkTimeDay today = WorkTimeDay.now();
     if (t != null) {
-      if(t.getWeekNumber() != DayEntry.INVALID_WEEK && MainApplication.getInstance().isDisplayWeek()) {
+      MainApplication app = (MainApplication) mContext.getApplicationContext();
+      if (t.getWeekNumber() != DayEntry.INVALID_WEEK && app.isDisplayWeek()) {
         holder.tvWeek.setMinHeight(0); // Min Height
         holder.tvWeek.setMinimumHeight(0); // Min Height
-        holder.tvWeek.setHeight(MainApplication.getInstance().getDayRowsHeight() - 30);
+        holder.tvWeek.setHeight(app.getDayRowsHeight() - 30);
         holder.tvWeek.setVisibility(View.VISIBLE);
         holder.tvDay.setVisibility(View.GONE);
         holder.tvStart.setVisibility(View.GONE);
@@ -124,7 +129,8 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
         holder.llDay.setBackgroundColor(mContext.getResources().getColor(R.color.color_week, null));
         String week = mContext.getString(R.string.week).toLowerCase();
         week = Character.toUpperCase(week.charAt(0)) + week.substring(1);
-        holder.tvWeek.setText((week + " " + t.getWeekNumber()));
+        String s = week + " " + t.getWeekNumber();
+        holder.tvWeek.setText(s);
       } else {
         setRowHeight(holder.tvDay);
         setRowHeight(holder.tvStart);
@@ -140,65 +146,69 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
         holder.tvTotal.setVisibility(View.VISIBLE);
         holder.tvOver.setVisibility(View.VISIBLE);
         Calendar cal = t.getDay().toCalendar();
-        int bg_morning;
-        int bg_afternoon;
+        int bgMorning;
+        int bgAfternoon;
         if (t.matchSimpleDate(today)) {
-          bg_afternoon = bg_morning = mContext.getResources().getColor(R.color.green, null);
+          bgAfternoon = bgMorning = mContext.getResources().getColor(R.color.green, null);
         } else {
           if (cal.get(Calendar.DAY_OF_WEEK) == Calendar.SATURDAY && t.getTypeMorning() != DayType.PUBLIC_HOLIDAY)
-            bg_morning = mContext.getResources().getColor(R.color.blue_1, null);
+            bgMorning = mContext.getResources().getColor(R.color.blue_1, null);
           else if (cal.get(Calendar.DAY_OF_WEEK) == Calendar.SUNDAY && t.getTypeMorning() != DayType.PUBLIC_HOLIDAY)
-            bg_morning = mContext.getResources().getColor(R.color.blue_2, null);
-          else
-            bg_morning = (t.getTypeMorning() == DayType.HOLIDAY) ?
-                mContext.getResources().getColor(R.color.purple, null) :
-                (t.getTypeMorning() == DayType.PUBLIC_HOLIDAY) ?
-                    mContext.getResources().getColor(R.color.purple2, null) :
-                    (t.getTypeMorning() == DayType.UNPAID) ?
-                        mContext.getResources().getColor(R.color.red, null) :
-                        (t.getTypeMorning() == DayType.SICKNESS) ?
-                            mContext.getResources().getColor(R.color.orange, null) :
-                            (t.getTypeMorning() == DayType.RECOVERY) ?
-                                mContext.getResources().getColor(R.color.gray, null) :
-                                mContext.getResources().getColor(android.R.color.transparent, null);
+            bgMorning = mContext.getResources().getColor(R.color.blue_2, null);
+          else {
+            if (t.getTypeMorning() == DayType.HOLIDAY)
+              bgMorning = mContext.getResources().getColor(R.color.purple, null);
+            else if (t.getTypeMorning() == DayType.PUBLIC_HOLIDAY)
+              bgMorning = mContext.getResources().getColor(R.color.purple2, null);
+            else if (t.getTypeMorning() == DayType.UNPAID)
+              bgMorning = mContext.getResources().getColor(R.color.red, null);
+            else if (t.getTypeMorning() == DayType.SICKNESS)
+              bgMorning = mContext.getResources().getColor(R.color.orange, null);
+            else if (t.getTypeMorning() == DayType.RECOVERY)
+              bgMorning = mContext.getResources().getColor(R.color.gray, null);
+            else
+              bgMorning = mContext.getResources().getColor(android.R.color.transparent, null);
+          }
 
           if (cal.get(Calendar.DAY_OF_WEEK) == Calendar.SATURDAY && t.getTypeAfternoon() != DayType.PUBLIC_HOLIDAY)
-            bg_afternoon = mContext.getResources().getColor(R.color.blue_1, null);
+            bgAfternoon = mContext.getResources().getColor(R.color.blue_1, null);
           else if (cal.get(Calendar.DAY_OF_WEEK) == Calendar.SUNDAY && t.getTypeAfternoon() != DayType.PUBLIC_HOLIDAY)
-            bg_afternoon = mContext.getResources().getColor(R.color.blue_2, null);
-          else
-            bg_afternoon = (t.getTypeAfternoon() == DayType.HOLIDAY) ?
-                mContext.getResources().getColor(R.color.purple, null) :
-                (t.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY) ?
-                    mContext.getResources().getColor(R.color.purple2, null) :
-                    (t.getTypeAfternoon() == DayType.UNPAID) ?
-                        mContext.getResources().getColor(R.color.red, null) :
-                        (t.getTypeAfternoon() == DayType.SICKNESS) ?
-                            mContext.getResources().getColor(R.color.orange, null) :
-                            (t.getTypeAfternoon() == DayType.RECOVERY) ?
-                                mContext.getResources().getColor(R.color.gray, null) :
-                                mContext.getResources().getColor(android.R.color.transparent, null);
+            bgAfternoon = mContext.getResources().getColor(R.color.blue_2, null);
+          else {
+            if (t.getTypeAfternoon() == DayType.HOLIDAY)
+              bgAfternoon = mContext.getResources().getColor(R.color.purple, null);
+            else if (t.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY)
+              bgAfternoon = mContext.getResources().getColor(R.color.purple2, null);
+            else if (t.getTypeAfternoon() == DayType.UNPAID)
+              bgAfternoon = mContext.getResources().getColor(R.color.red, null);
+            else if (t.getTypeAfternoon() == DayType.SICKNESS)
+              bgAfternoon = mContext.getResources().getColor(R.color.orange, null);
+            else if (t.getTypeAfternoon() == DayType.RECOVERY)
+              bgAfternoon = mContext.getResources().getColor(R.color.gray, null);
+            else
+              bgAfternoon = mContext.getResources().getColor(android.R.color.transparent, null);
+          }
         }
-        if (bg_morning == bg_afternoon) {
-          holder.llDay.setBackgroundColor(bg_morning);
+        if (bgMorning == bgAfternoon) {
+          holder.llDay.setBackgroundColor(bgMorning);
         } else {
-          UIHelper.applyLinearGradient(holder.llDay, bg_afternoon, bg_morning);
+          UIHelper.applyLinearGradient(holder.llDay, bgAfternoon, bgMorning);
         }
 
 
         /* Update the texts */
         if (holder.tvDay != null) {
-          holder.tvDay.setText((String.format(Locale.US, "%02d", cal.get(Calendar.DAY_OF_MONTH)) + " " +
-              DayEntry.getDayString2lt(mContext, cal.get(Calendar.DAY_OF_WEEK))));
+          String s = String.format(Locale.US, "%02d", cal.get(Calendar.DAY_OF_MONTH)) + " " +
+            DayEntry.getDayString2lt(mContext, cal.get(Calendar.DAY_OF_WEEK));
+          holder.tvDay.setText(s);
         }
         if (holder.tvStart != null) {
           if (isPaidButNotWorkedMorning(t)) {
-            if(!isNotValidAfternoon(t))
+            if (!isNotValidAfternoon(t))
               holder.tvStart.setText(t.getStartAfternoon().timeString());
             else
               holder.tvStart.setText("*");
-          }
-          else if (t.getTypeMorning() == DayType.RECOVERY && !isNotValidAfternoon(t))
+          } else if (t.getTypeMorning() == DayType.RECOVERY && !isNotValidAfternoon(t))
             holder.tvStart.setText(t.getStartAfternoon().timeString());
           else if (isNotValidMorning(t))
             holder.tvStart.setText("-");
@@ -209,16 +219,15 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
         if (holder.tvEnd != null) {
           if (isPaidButNotWorkedAfternoon(t)) {
             holder.tvEnd.setText("*");
-            if(!isNotValidMorning(t))
+            if (!isNotValidMorning(t))
               holder.tvEnd.setText(t.getEndMorning().timeString());
             else
               holder.tvEnd.setText("*");
-          }
-          else if (t.getTypeAfternoon() == DayType.RECOVERY && !isNotValidMorning(t))
+          } else if (t.getTypeAfternoon() == DayType.RECOVERY && !isNotValidMorning(t))
             holder.tvEnd.setText(t.getEndMorning().timeString());
           else if (isNotValidAfternoon(t)) {
             WorkTimeDay w = t.getEndMorning();
-            if (w.timeString().equals("00:00") || isNotValidMorning(t))
+            if (w.timeString().equals(TIME_ZERO) || isNotValidMorning(t))
               holder.tvEnd.setText("-");
             else
               holder.tvEnd.setText(t.getEndMorning().timeString());
@@ -231,7 +240,7 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
             holder.tvPause.setText("*");
           else {
             WorkTimeDay w = t.getPause();
-            if (isNotValidMorning(t) && isNotValidAfternoon(t) || w.timeString().equals("00:00"))
+            if (isNotValidMorning(t) && isNotValidAfternoon(t) || w.timeString().equals(TIME_ZERO))
               holder.tvPause.setText("-");
             else {
               holder.tvPause.setText(t.getPause().timeString());
@@ -242,7 +251,7 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
         if (holder.tvTotal != null) {
           if (isPaidButNotWorkedMorning(t) && isPaidButNotWorkedAfternoon(t))
             holder.tvTotal.setText("*");
-          else if (isNotValidMorning(t) && isNotValidAfternoon(t) || w.timeString().equals("00:00"))
+          else if (isNotValidMorning(t) && isNotValidAfternoon(t) || w.timeString().equals(TIME_ZERO))
             holder.tvTotal.setText("-");
           else {
             holder.tvTotal.setText(t.getWorkTime().timeString());
@@ -254,7 +263,7 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
             holder.tvOver.setTypeface(Typeface.MONOSPACE, Typeface.NORMAL);
             holder.tvOver.setGravity(Gravity.CENTER);
             holder.tvOver.setTextColor(holder.tvOverColors);
-          } else if (isNotValidMorning(t) && isNotValidAfternoon(t) || w.timeString().equals("00:00")) {
+          } else if (isNotValidMorning(t) && isNotValidAfternoon(t) || w.timeString().equals(TIME_ZERO)) {
             holder.tvOver.setText("-");
             holder.tvOver.setTypeface(Typeface.MONOSPACE, Typeface.NORMAL);
             holder.tvOver.setGravity(Gravity.CENTER);
@@ -286,55 +295,59 @@ public class DaysEntriesArrayAdapter extends ArrayAdapter<DayEntry> {
 
   /**
    * Sets the row height.
+   *
    * @param v The current TextView.
    */
   private void setRowHeight(TextView v) {
     v.setMinHeight(0); // Min Height
     v.setMinimumHeight(0); // Min Height
-    v.setHeight(MainApplication.getInstance().getDayRowsHeight()); // Height
+    v.setHeight(((MainApplication) mContext.getApplicationContext()).getDayRowsHeight()); // Height
   }
 
   /**
    * Tests if the morning value of the day entry is valid or not.
+   *
    * @param t The day entry to test.
    * @return boolean
    */
   private boolean isNotValidMorning(DayEntry t) {
     return (t.getTypeMorning() != DayType.AT_WORK || (!t.getStartMorning().isValidTime()
-        && !t.getEndMorning().isValidTime()));
+      && !t.getEndMorning().isValidTime()));
   }
 
   /**
    * Tests if the afternoon value of the day entry is valid or not.
+   *
    * @param t The day entry to test.
    * @return boolean
    */
   private boolean isNotValidAfternoon(DayEntry t) {
     return (t.getTypeAfternoon() != DayType.AT_WORK || (!t.getStartAfternoon().isValidTime()
-        && !t.getEndAfternoon().isValidTime()));
+      && !t.getEndAfternoon().isValidTime()));
   }
-
 
 
   /**
    * Tests if the morning value of the day entry is paid but not worked or not.
+   *
    * @param t The day entry to test.
    * @return boolean
    */
   private boolean isPaidButNotWorkedMorning(DayEntry t) {
     return t.getTypeMorning() == DayType.PUBLIC_HOLIDAY || ((t.getTypeMorning() == DayType.HOLIDAY
-        || t.getTypeMorning() == DayType.SICKNESS) && t.getStartMorning().isValidTime()
-        && t.getEndMorning().isValidTime());
+      || t.getTypeMorning() == DayType.SICKNESS) && t.getStartMorning().isValidTime()
+      && t.getEndMorning().isValidTime());
   }
 
   /**
    * Tests if the afternoon value of the day entry is paid but not worked or not.
+   *
    * @param t The day entry to test.
    * @return boolean
    */
   private boolean isPaidButNotWorkedAfternoon(DayEntry t) {
     return t.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY || ((t.getTypeAfternoon() == DayType.HOLIDAY
-        || t.getTypeAfternoon() == DayType.SICKNESS) && (t.getStartAfternoon().isValidTime()
-        && t.getEndAfternoon().isValidTime()));
+      || t.getTypeAfternoon() == DayType.SICKNESS) && (t.getStartAfternoon().isValidTime()
+      && t.getEndAfternoon().isValidTime()));
   }
 }
