@@ -1,25 +1,26 @@
 package fr.ralala.worktime.ui.changelog;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Manage the changelog
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class ChangeLogIds {
-  private int mStringBackgroundColor;
-  private int mStringChangelogTitle;
-  private int mStringChangelogFullTitle;
-  private int mStringChangelogShowFull;
-  private int mStringChangelogOkButton;
-  private int mRawChangelog;
+  private final int mStringBackgroundColor;
+  private final int mStringChangelogTitle;
+  private final int mStringChangelogFullTitle;
+  private final int mStringChangelogShowFull;
+  private final int mStringChangelogOkButton;
+  private final int mRawChangelog;
 
   public ChangeLogIds(final int rawChangelog, final int stringChangelogOkButton, final int stringBackgroundColor,
-      final int stringChangelogTitle, final int stringChangelogFullTitle,
-      final int stringChangelogShowFull) {
+                      final int stringChangelogTitle, final int stringChangelogFullTitle,
+                      final int stringChangelogShowFull) {
     mStringChangelogOkButton = stringChangelogOkButton;
     mStringBackgroundColor = stringBackgroundColor;
     mStringChangelogTitle = stringChangelogTitle;
@@ -30,6 +31,7 @@ public class ChangeLogIds {
 
   /**
    * Returns the String resource for the label of the OK button.
+   *
    * @return int
    */
   int getStringChangelogOkButton() {
@@ -38,6 +40,7 @@ public class ChangeLogIds {
 
   /**
    * Returns the String resource for the background.
+   *
    * @return int
    */
   int getStringBackgroundColor() {
@@ -46,6 +49,7 @@ public class ChangeLogIds {
 
   /**
    * Returns the String resource for the title.
+   *
    * @return int
    */
   int getStringChangelogTitle() {
@@ -54,6 +58,7 @@ public class ChangeLogIds {
 
   /**
    * Returns the String resource for the full title.
+   *
    * @return int
    */
   int getStringChangelogFullTitle() {
@@ -62,6 +67,7 @@ public class ChangeLogIds {
 
   /**
    * Returns the String resource for show full field.
+   *
    * @return int
    */
   int getStringChangelogShowFull() {
@@ -71,6 +77,7 @@ public class ChangeLogIds {
 
   /**
    * Returns the resource id of the changelog file.
+   *
    * @return int
    */
   int getRawChangelog() {
