@@ -4,11 +4,13 @@ import android.app.Activity;
 import android.content.DialogInterface;
 import android.content.res.Resources;
 import android.graphics.Typeface;
-import android.support.v7.app.AlertDialog;
 import android.util.SparseArray;
 import android.view.Gravity;
+import android.view.ViewGroup;
 import android.widget.GridLayout;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AlertDialog;
 
 import java.util.ArrayList;
 import java.util.Calendar;
@@ -25,21 +27,22 @@ import fr.ralala.worktime.models.WorkTimeDay;
 import fr.ralala.worktime.utils.AndroidHelper;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Manage the dialog box containing the month details
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class MonthDetailsDialog implements DialogInterface.OnClickListener {
 
-  private Activity mActivity;
-  private MainApplication mApp;
+  private final Activity mActivity;
+  private final MainApplication mApp;
   private AlertDialog mAlertDialog = null;
 
-  private class Item {
+  private static class Item {
     List<DayEntry> wDays = new ArrayList<>();
     WorkTimeDay w = new WorkTimeDay();
     WorkTimeDay rec = new WorkTimeDay();
@@ -49,8 +52,9 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
 
   /**
    * Creates the dialog used to display the details of a month.
+   *
    * @param activity The main activity.
-   * @param app The main application.
+   * @param app      The main application.
    */
   public MonthDetailsDialog(final Activity activity, final MainApplication app) {
     mActivity = activity;
@@ -59,8 +63,9 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
 
   /**
    * Reloads the details.
+   *
    * @param month The selected month.
-   * @param year The selected year.
+   * @param year  The selected year.
    */
   public void reloadDetails(int month, int year) {
     final String currency = mApp.getCurrency();
@@ -75,9 +80,9 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
       r.getDimensionPixelOffset(R.dimen.activity_horizontal_margin),
       r.getDimensionPixelOffset(R.dimen.activity_vertical_margin));
 
-    GridLayout.LayoutParams param =new GridLayout.LayoutParams();
-    param.height = GridLayout.LayoutParams.MATCH_PARENT;
-    param.width = GridLayout.LayoutParams.MATCH_PARENT;
+    GridLayout.LayoutParams param = new GridLayout.LayoutParams();
+    param.height = ViewGroup.LayoutParams.MATCH_PARENT;
+    param.width = ViewGroup.LayoutParams.MATCH_PARENT;
     gl.setLayoutParams(param);
     dialogBuilder.setView(gl);
     WorkTimeDay wtdTotalWorkTime = new WorkTimeDay();
@@ -93,9 +98,9 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
     ctime.set(Calendar.MONTH, month);
     List<DayEntry> days = mApp.getDaysFactory().list(year, month + 1, -1);
     SparseArray<Item> weeks = new SparseArray<>();
-    for(DayEntry de : days) {
+    for (DayEntry de : days) {
       ctime.set(Calendar.DAY_OF_MONTH, de.getDay().getDay());
-      if(de.getTypeMorning() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.AT_WORK) {
+      if (de.getTypeMorning() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.AT_WORK) {
         Item i;
         if (AndroidHelper.notContainsKey(weeks, ctime.get(Calendar.WEEK_OF_YEAR)))
           weeks.put(ctime.get(Calendar.WEEK_OF_YEAR), new Item());
@@ -108,15 +113,15 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
       }
     }
     List<Integer> keys = new ArrayList<>();
-    for(int i = 0; i < weeks.size(); i++)
+    for (int i = 0; i < weeks.size(); i++)
       keys.add(weeks.keyAt(i));
     keys.sort(Comparator.naturalOrder());
-    for(int idx = 0; idx < keys.size();++idx, ++row) {
+    for (int idx = 0; idx < keys.size(); ++idx, ++row) {
       Integer w = keys.get(idx);
       Item i = weeks.get(w);
       WorkTimeDay wtdEstimatedHours = mApp.getEstimatedHours(i.wDays);
-      WorkTimeDay wtdOver = i.w.clone().delTime(wtdEstimatedHours);
-      WorkTimeDay wtdRec = i.rec.clone();
+      WorkTimeDay wtdOver = i.w.copy().delTime(wtdEstimatedHours);
+      WorkTimeDay wtdRec = i.rec.copy();
       totalWage += i.wage;
       addRow(r, gl, row, w,
         i.w.timeString(),
@@ -129,29 +134,30 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
 
     /* attach the listeners and init the default values */
     dialogBuilder.setPositiveButton(R.string.ok, this);
-    if(mAlertDialog != null && mAlertDialog.isShowing()) mAlertDialog.dismiss();
+    if (mAlertDialog != null && mAlertDialog.isShowing()) mAlertDialog.dismiss();
     mAlertDialog = dialogBuilder.create();
   }
 
   /**
    * Adds a new row.
-   * @param r The application resource.
-   * @param gl GridLayout container.
-   * @param row The current row.
-   * @param week The current week.
-   * @param wt The work time value.
-   * @param ot The over time value.
-   * @param rec The recovery time value.
-   * @param wage The wage value.
+   *
+   * @param r        The application resource.
+   * @param gl       GridLayout container.
+   * @param row      The current row.
+   * @param week     The current week.
+   * @param wt       The work time value.
+   * @param ot       The over time value.
+   * @param rec      The recovery time value.
+   * @param wage     The wage value.
    * @param currency The currency.
    */
   private void addRow(Resources r, GridLayout gl, int row, int week, String wt, String ot, String rec, double wage, String currency) {
     GridLayout.LayoutParams param;
-    for(int j = 0; j < 9; ++j) {
+    for (int j = 0; j < 9; ++j) {
       TextView tvWeek = new TextView(mActivity);
-      param =new GridLayout.LayoutParams();
-      param.height = GridLayout.LayoutParams.WRAP_CONTENT;
-      param.width = GridLayout.LayoutParams.WRAP_CONTENT;
+      param = new GridLayout.LayoutParams();
+      param.height = ViewGroup.LayoutParams.WRAP_CONTENT;
+      param.width = ViewGroup.LayoutParams.WRAP_CONTENT;
       param.rightMargin = 20;
       param.topMargin = 5;
       param.setGravity(Gravity.START);
@@ -159,33 +165,41 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
       param.rowSpec = GridLayout.spec(row);
       param.setGravity(Gravity.FILL);
       tvWeek.setLayoutParams(param);
-      if(j == 0 || week == -1) tvWeek.setTypeface(tvWeek.getTypeface(), Typeface.BOLD);
-      if(j != 0)
+      if (j == 0 || week == -1) tvWeek.setTypeface(tvWeek.getTypeface(), Typeface.BOLD);
+      if (j != 0)
         tvWeek.setGravity(Gravity.END);
 
       switch (j) {
-        case 0:
-          if(week == -1) {
-            String s = r.getString(R.string.total) + ":";
-            tvWeek.setText(s);
-          } else
-            tvWeek.setText((r.getString(R.string.week_letter) + " " + String.format(Locale.US, "%02d", week) + ":"));
+        case 0: {
+          String s;
+          if (week == -1) {
+            s = r.getString(R.string.total) + ":";
+          } else {
+            s = r.getString(R.string.week_letter);
+            s += " " + String.format(Locale.US, "%02d", week) + ":";
+          }
+          tvWeek.setText(s);
           break;
+        }
         case 1:
           tvWeek.setText(wt);
           break;
-        case 2:
-          tvWeek.setText(("(" + mActivity.getString(R.string.overtime_min) + ": "));
+        case 2: {
+          String s = "(" + mActivity.getString(R.string.overtime_min) + ": ";
+          tvWeek.setText(s);
           break;
+        }
         case 3:
           tvWeek.setText(ot);
           break;
         case 4:
           tvWeek.setText(",");
           break;
-        case 5:
-          tvWeek.setText((mActivity.getString(R.string.recovery_min) + ": "));
+        case 5: {
+          String s = mActivity.getString(R.string.recovery_min) + ": ";
+          tvWeek.setText(s);
           break;
+        }
         case 6:
           tvWeek.setText(rec);
           break;
@@ -211,7 +225,8 @@ public class MonthDetailsDialog implements DialogInterface.OnClickListener {
 
   /**
    * Called when the user click on the OK button.
-   * @param dialog Not used.
+   *
+   * @param dialog      Not used.
    * @param whichButton Not used.
    */
   public void onClick(DialogInterface dialog, int whichButton) {
