@@ -4,8 +4,6 @@ package fr.ralala.worktime.ui.fragments;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -15,6 +13,9 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.ListView;
 import android.widget.Spinner;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 
 import org.apache.poi.ss.usermodel.Sheet;
 
@@ -26,27 +27,28 @@ import java.util.List;
 import java.util.Locale;
 import java.util.TimeZone;
 
-import fr.ralala.worktime.models.DayType;
-import fr.ralala.worktime.ui.activities.MainActivity;
-import fr.ralala.worktime.utils.AndroidHelper;
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
-import fr.ralala.worktime.ui.adapters.ExportListViewArrayAdapter;
 import fr.ralala.worktime.models.DayEntry;
+import fr.ralala.worktime.models.DayType;
 import fr.ralala.worktime.models.WorkTimeDay;
-import fr.ralala.worktime.utils.ExcelHelper;
+import fr.ralala.worktime.ui.activities.MainActivity;
+import fr.ralala.worktime.ui.adapters.ExportListViewArrayAdapter;
 import fr.ralala.worktime.ui.utils.UIHelper;
+import fr.ralala.worktime.utils.AndroidHelper;
+import fr.ralala.worktime.utils.ExcelHelper;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Management of the export fragment view
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
-public class ExportFragment extends Fragment implements AdapterView.OnItemSelectedListener, View.OnClickListener{
+public class ExportFragment extends Fragment implements AdapterView.OnItemSelectedListener, View.OnClickListener {
 
   private ExportListViewArrayAdapter mLvAdapter = null;
   private Spinner mSpinner = null;
@@ -55,8 +57,9 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
 
   /**
    * Called when the fragment is created.
-   * @param inflater The fragment inflater.
-   * @param container The fragment container.
+   *
+   * @param inflater           The fragment inflater.
+   * @param container          The fragment container.
    * @param savedInstanceState The saved instance state.
    * @return The created view.
    */
@@ -64,9 +67,9 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
   public View onCreateView(@NonNull final LayoutInflater inflater,
                            final ViewGroup container, final Bundle savedInstanceState) {
     final ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.fragment_export, container, false);
-    mActivity = (MainActivity)getActivity();
+    mActivity = (MainActivity) getActivity();
     assert mActivity != null;
-    mApp = MainApplication.getInstance();
+    mApp = (MainApplication) mActivity.getApplication();
     Button export = rootView.findViewById(R.id.btExport);
     export.setOnClickListener(this);
 
@@ -85,56 +88,58 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
     c.setTimeZone(TimeZone.getTimeZone("GMT"));
     c.setTime(new Date());
     List<Integer> years = mApp.getDaysFactory().getYears();
-    for(int i = 0; i < years.size(); i++) {
+    for (int i = 0; i < years.size(); i++) {
       Integer y = years.get(i);
       spAdapter.add(String.valueOf(y));
-      if(y == c.get(Calendar.YEAR)) mSpinner.setSelection(i);
+      if (y == c.get(Calendar.YEAR)) mSpinner.setSelection(i);
     }
     mSpinner.setOnItemSelectedListener(this);
-    //progressReload(Integer.parseInt(mSpinner.getSelectedItem().toString()));
     return rootView;
   }
 
   /**
    * Called when the user click on a button (export).
+   *
    * @param v The clicked view.
    */
   @Override
   public void onClick(View v) {
     final List<ExportListViewArrayAdapter.ExportEntry> entries = mLvAdapter.getCheckedItems();
-    if(entries.isEmpty()) {
+    if (entries.isEmpty()) {
       UIHelper.snack(mActivity, getString(R.string.export_no_items));
       return;
     }
     final String email = mApp.getEMail();
-    if(mApp.isExportMailEnabled() && email.isEmpty()) {
+    if (mApp.isExportMailEnabled() && email.isEmpty()) {
       UIHelper.snack(mActivity, getString(R.string.export_email_not_set));
       return;
     }
 
     final File file = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS),
       getString(R.string.app_name) + "_" + entries.get(0).year + ".xls");
-    if(file.exists()) {
+    if (file.exists()) {
       UIHelper.showConfirmDialog(mActivity, getString(R.string.confirm),
         getString(R.string.the_file_exists_part_1) + " " + file.getName() + " " + getString(R.string.the_file_exists_part_2),
-        (view) -> exportXLS(entries, email, file));
+        view -> exportXLS(entries, email, file));
     } else
       exportXLS(entries, email, file);
   }
 
   /**
    * Exports the selected entries in a XLS file.
+   *
    * @param entries The entries to export.
-   * @param email The email address used to send the file.
-   * @param file The file name.
+   * @param email   The email address used to send the file.
+   * @param file    The file name.
    */
   private void exportXLS(List<ExportListViewArrayAdapter.ExportEntry> entries, String email, File file) {
     try {
       ExcelHelper excel = new ExcelHelper(file);
-      for(int idx = 0; idx < entries.size(); ++idx) {
+      for (int idx = 0; idx < entries.size(); ++idx) {
         ExportListViewArrayAdapter.ExportEntry ee = entries.get(idx);
         Sheet sheet = excel.createSheet(AndroidHelper.getMonthString(ee.month));
-        int row = 0, column = 0;
+        int row = 0;
+        int column = 0;
 
         List<String> headers = new ArrayList<>();
         headers.add(getString(R.string.date).replaceAll(" :", "")); // A -> 65
@@ -144,33 +149,32 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
         headers.add(getString(R.string.hour_out)); // E -> 69
         headers.add(getString(R.string.break_time)); // F -> 70
         headers.add(getString(R.string.overtime)); // G -> 71
-        if(!mApp.isExportHideWage())
+        if (!mApp.isExportHideWage())
           headers.add(getString(R.string.wage).replaceAll(" :", "")); // H -> 72
         excel.createHorizontalHeader(sheet, row, column, headers.toArray(new String[]{}));
         List<DayEntry> works = mApp.getDaysFactory().list(ee.year, ee.month + 1, -1);
-        column = 0;
         row++;
         for (DayEntry de : works) {
           excel.addLabel(sheet, row, column, de.getDay().dateString(), false);
-          excel.addLabel(sheet, row, column+1, de.getTypeMorning() != DayType.AT_WORK ? de.getTypeMorning().string(mActivity) : "", false);
-          excel.addLabel(sheet, row, column+2, de.getTypeAfternoon() != DayType.AT_WORK ? de.getTypeAfternoon().string(mActivity) : "", false);
+          excel.addLabel(sheet, row, column + 1, de.getTypeMorning() != DayType.AT_WORK ? de.getTypeMorning().string(mActivity) : "", false);
+          excel.addLabel(sheet, row, column + 2, de.getTypeAfternoon() != DayType.AT_WORK ? de.getTypeAfternoon().string(mActivity) : "", false);
           double wage = .0f;
-          int column_offset= 3;
-          if(de.getTypeMorning() != DayType.AT_WORK && de.getTypeAfternoon() != DayType.AT_WORK) {
-            excel.addTime(sheet, row, column + (column_offset++), WorkTimeDay.timeString(0, 0), false);
-            excel.addTime(sheet, row, column + (column_offset++), WorkTimeDay.timeString(0, 0), false);
-            excel.addTime(sheet, row, column + (column_offset++), WorkTimeDay.timeString(0, 0), false);
-            excel.addTime(sheet, row, column + (column_offset++), WorkTimeDay.timeString(0, 0), false);
+          int columnOffset = 3;
+          if (de.getTypeMorning() != DayType.AT_WORK && de.getTypeAfternoon() != DayType.AT_WORK) {
+            excel.addTime(sheet, row, column + (columnOffset++), WorkTimeDay.timeString(0, 0), false);
+            excel.addTime(sheet, row, column + (columnOffset++), WorkTimeDay.timeString(0, 0), false);
+            excel.addTime(sheet, row, column + (columnOffset++), WorkTimeDay.timeString(0, 0), false);
+            excel.addTime(sheet, row, column + (columnOffset++), WorkTimeDay.timeString(0, 0), false);
           } else {
-            excel.addTime(sheet, row, column + (column_offset++), de.getStartMorning().timeString(), false);
-            excel.addTime(sheet, row, column + (column_offset++), de.getEndAfternoon().timeString(), false);
-            excel.addTime(sheet, row, column + (column_offset++), de.getPause().timeString(), false);
-            excel.addTime(sheet, row, column + (column_offset++), de.getOverTime().timeString(), false);
+            excel.addTime(sheet, row, column + (columnOffset++), de.getStartMorning().timeString(), false);
+            excel.addTime(sheet, row, column + (columnOffset++), de.getEndAfternoon().timeString(), false);
+            excel.addTime(sheet, row, column + (columnOffset++), de.getPause().timeString(), false);
+            excel.addTime(sheet, row, column + (columnOffset++), de.getOverTime().timeString(), false);
             if (!mApp.isExportHideWage())
               wage = de.getWorkTimePay(mApp.getAmountByHour());
           }
           if (!mApp.isExportHideWage())
-            excel.addNumber(sheet, row++, column + (column_offset), wage);
+            excel.addNumber(sheet, row++, column + (columnOffset), wage);
         }
         column = 4;
         excel.addLabel(sheet, row, column++, getString(R.string.total), true);
@@ -185,7 +189,7 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
 
       excel.write();
       Uri uri = Uri.fromFile(file);
-      if(mApp.isExportMailEnabled()) {
+      if (mApp.isExportMailEnabled()) {
         AndroidHelper.sentMailTo(mActivity, email, uri,
           getString(R.string.export_email_subject),
           getString(R.string.export_email_body),
@@ -193,8 +197,7 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
         UIHelper.snack(mActivity, getString(R.string.email_sent_to_with_mail) + " " + email);
       } else
         UIHelper.snack(mActivity, getString(R.string.email_sent_to_without_mail));
-      //if(file.exists()) file.delete();
-    } catch(Exception e) {
+    } catch (Exception e) {
       Log.e(getClass().getSimpleName(), "Exception: " + e.getMessage(), e);
       UIHelper.snack(mActivity, getString(R.string.error) + ": " + e.getMessage());
     }
@@ -202,10 +205,11 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
 
   /**
    * Called when the user click on the dropdown list.
-   * @param parentView See official javadoc.
+   *
+   * @param parentView       See official javadoc.
    * @param selectedItemView See official javadoc.
-   * @param position See official javadoc.
-   * @param id See official javadoc.
+   * @param position         See official javadoc.
+   * @param id               See official javadoc.
    */
   @Override
   public void onItemSelected(AdapterView<?> parentView, View selectedItemView, int position, long id) {
@@ -214,14 +218,17 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
 
   /**
    * See official javadoc.
+   *
    * @param parentView See official javadoc.
    */
   @Override
   public void onNothingSelected(AdapterView<?> parentView) {
+    /* Nothing */
   }
 
   /**
    * Reloads the available month associated to the desired year.
+   *
    * @param year The desired year.
    */
   private void progressReload(int year) {
@@ -231,25 +238,27 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
       mActivity.runOnUiThread(() -> mActivity.progressDismiss());
     }).start();
   }
+
   /**
    * Reloads the available month associated to the desired year.
+   *
    * @param year The desired year.
    */
   private void reload(int year) {
     mActivity.runOnUiThread(() -> mLvAdapter.clear());
     Locale locale = getResources().getConfiguration().getLocales().get(0);
-    for(int i = 0; i < 12; ++i) {
+    for (int i = 0; i < 12; ++i) {
 
       WorkTimeDay total = new WorkTimeDay();
       double pay = 0;
       List<DayEntry> days = mApp.getDaysFactory().list(year, i + 1, -1);
-      for(DayEntry de : days) {
-        if(de != null && (de.getTypeMorning() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.AT_WORK)) {
+      for (DayEntry de : days) {
+        if (de != null && (de.getTypeMorning() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.AT_WORK)) {
           pay += de.getWorkTimePay(mApp.getAmountByHour());
           total.addTime(de.getWorkTime());
         }
       }
-      if(!total.isValidTime()) continue;
+      if (!total.isValidTime()) continue;
       String info = !mApp.isExportHideWage() ?
         String.format(locale, "%02d:%02d %s (%.02f %s)",
           total.getHours(), total.getMinutes(),
@@ -258,7 +267,7 @@ public class ExportFragment extends Fragment implements AdapterView.OnItemSelect
         :
         String.format(locale, "%02d:%02d %s",
           total.getHours(), total.getMinutes(),
-          getString(R.string.hours_lower_case)) ;
+          getString(R.string.hours_lower_case));
       ExportListViewArrayAdapter.ExportEntry ee = new ExportListViewArrayAdapter.ExportEntry();
       ee.text = AndroidHelper.getMonthString(i);
       ee.info = info;
