@@ -2,13 +2,7 @@ package fr.ralala.worktime.ui.fragments;
 
 
 import android.app.Activity;
-import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -16,13 +10,24 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
-import java.util.List;
+import androidx.activity.result.ActivityResult;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.view.MenuHost;
+import androidx.core.view.MenuProvider;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Lifecycle;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
-import fr.ralala.worktime.models.ProfileEntry;
-import fr.ralala.worktime.ui.activities.DayActivity;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
+import java.util.List;
 
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
+import fr.ralala.worktime.launchers.LauncherCallback;
+import fr.ralala.worktime.models.ProfileEntry;
 import fr.ralala.worktime.ui.activities.MainActivity;
 import fr.ralala.worktime.ui.adapters.EntriesArrayAdapter;
 import fr.ralala.worktime.ui.adapters.ProfilesEntriesArrayAdapter;
@@ -32,17 +37,18 @@ import fr.ralala.worktime.utils.AndroidHelper;
 
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Management of the profiles fragment view
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
-public class ProfileFragment  extends Fragment implements View.OnClickListener,
-    SwipeEditDeleteRecyclerViewItem.SwipeEditDeleteRecyclerViewItemListener,
-    EntriesArrayAdapter.OnLongPressListener<ProfileEntry> {
+public class ProfileFragment extends Fragment implements View.OnClickListener,
+  SwipeEditDeleteRecyclerViewItem.SwipeEditDeleteRecyclerViewItemListener,
+  EntriesArrayAdapter.OnLongPressListener<ProfileEntry>, LauncherCallback {
 
   private ProfilesEntriesArrayAdapter mAdapter = null;
   private MainApplication mApp = null;
@@ -51,18 +57,9 @@ public class ProfileFragment  extends Fragment implements View.OnClickListener,
 
   /**
    * Called when the fragment is created.
-   * @param savedInstanceState The saved instance state.
-   */
-  @Override
-  public void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setHasOptionsMenu(true);
-  }
-
-  /**
-   * Called when the fragment is created.
-   * @param inflater The fragment inflater.
-   * @param container The fragment container.
+   *
+   * @param inflater           The fragment inflater.
+   * @param container          The fragment container.
    * @param savedInstanceState The saved instance state.
    * @return The created view.
    */
@@ -70,11 +67,11 @@ public class ProfileFragment  extends Fragment implements View.OnClickListener,
   public View onCreateView(@NonNull final LayoutInflater inflater,
                            final ViewGroup container, final Bundle savedInstanceState) {
     final ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.fragment_profile, container, false);
-    mActivity = (MainActivity)getActivity();
+    mActivity = (MainActivity) getActivity();
     assert mActivity != null;
     FloatingActionButton fab = rootView.findViewById(R.id.fab);
     fab.setOnClickListener(this);
-    mApp = MainApplication.getInstance();
+    mApp = (MainApplication) mActivity.getApplication();
 
     mRecyclerView = rootView.findViewById(R.id.profiles);
     mRecyclerView.setHasFixedSize(true);
@@ -86,6 +83,30 @@ public class ProfileFragment  extends Fragment implements View.OnClickListener,
     return rootView;
   }
 
+  @Override
+  public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+    MenuHost menuHost = mActivity;
+    // Add menu items without using the Fragment Menu APIs
+    // Note how we can tie the MenuProvider to the viewLifecycleOwner
+    // and an optional Lifecycle.State (here, RESUMED) to indicate when
+    // the menu should be visible
+    menuHost.addMenuProvider(new MenuProvider() {
+      public void onCreateMenu(@NonNull Menu menu, @NonNull MenuInflater menuInflater) {
+        // Add menu items here
+        menuInflater.inflate(R.menu.profile_fragment, menu);
+      }
+
+      public boolean onMenuItemSelected(@NonNull MenuItem menuItem) {
+        if (menuItem.getItemId() == R.id.action_help) {
+          Activity a = getActivity();
+          if (a != null)
+            UIHelper.snack(a, getString(R.string.help_profile));
+        }
+        return true;
+      }
+    }, getViewLifecycleOwner(), Lifecycle.State.RESUMED);
+  }
+
   /**
    * Refresh the adapter view.
    */
@@ -95,7 +116,7 @@ public class ProfileFragment  extends Fragment implements View.OnClickListener,
       final List<ProfileEntry> list = mApp.getProfilesFactory().list();
       mActivity.runOnUiThread(() -> {
         mAdapter = new ProfilesEntriesArrayAdapter(mRecyclerView,
-            getContext(), R.layout.listview_item, list);
+          getContext(), R.layout.listview_item, list);
         mAdapter.setOnLongPressListener(this);
         mRecyclerView.setAdapter(mAdapter);
         mAdapter.safeNotifyDataSetChanged();
@@ -106,96 +127,65 @@ public class ProfileFragment  extends Fragment implements View.OnClickListener,
 
   /**
    * Called when a long click is captured.
+   *
    * @param t The associated item.
    */
   @Override
   public void onLongPressListener(@NonNull ProfileEntry t) {
-    if(!Double.isNaN(t.getLongitude()) && !Double.isNaN(t.getLatitude())) {
-      mActivity.getVibrator().vibrate(50);
+    if (!Double.isNaN(t.getLongitude()) && !Double.isNaN(t.getLatitude())) {
+      AndroidHelper.vibrate(mActivity, 50);
       AndroidHelper.openDefaultNavigationApp(mActivity, t.getName(), t.getLocation());
     } else
       UIHelper.toast(mActivity, (getString(R.string.no_location_p1) + t.getName() + getString(R.string.no_location_p2)));
   }
 
-  /**
-   * Called when the options menu is created.
-   * @param menu The menu to inflate.
-   * @param inflater The menu inflater.
-   */
   @Override
-  public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-    inflater.inflate(R.menu.profile_fragment, menu);
-    super.onCreateOptionsMenu(menu, inflater);
-  }
-
-  /**
-   * Called when a menu is selected.
-   * @param item The selected item.
-   * @return boolean
-   */
-  @Override
-  public boolean onOptionsItemSelected(MenuItem item) {
-    switch (item.getItemId()) {
-      case R.id.action_help:
-        Activity a = getActivity();
-        if(a == null)
-          break;
-        UIHelper.snack(a, getString(R.string.help_profile));
-        break;
-    }
-    return true;
-  }
-
-  /**
-   * Receive the result from a previous call to startActivityForResult
-   * @param requestCode The integer request code originally supplied to startActivityForResult.
-   * @param resultCode The integer result code returned by the child activity through its setResult().
-   * @param data An Intent, which can return result data to the caller (various data can be attached to Intent "extras").
-   */
-  @Override
-  public void onActivityResult(int requestCode, int resultCode, Intent data) {
-    if(requestCode == DayActivity.REQUEST_START_ACTIVITY)
-      refreshView();
+  public void onLauncherResult(ActivityResult result) {
+    refreshView();
   }
 
   /**
    * Called when the user click on a button (fab).
+   *
    * @param view The clicked view.
    */
   @Override
   public void onClick(View view) {
-    DayActivity.startActivity(this, "null", false);
+    mActivity.getLauncherDayActivity().startActivity(this, "null", false);
   }
+
   /**
    * Called when a ViewHolder is swiped from left to right by the user.
+   *
    * @param adapterPosition The position in the adapter.
    */
   @Override
   public void onClickEdit(int adapterPosition) {
     ProfileEntry pe = mAdapter.getItem(adapterPosition);
-    if(pe == null) return;
-    DayActivity.startActivity(this, pe.getName(), false);
+    if (pe == null) return;
+    mActivity.getLauncherDayActivity().startActivity(this, pe.getName(), false);
   }
 
   /**
    * Called when a ViewHolder is swiped from right to left by the user.
+   *
    * @param adapterPosition The position in the adapter.
    */
   @Override
   public void onClickDelete(int adapterPosition) {
     ProfileEntry pe = mAdapter.getItem(adapterPosition);
-    if(pe == null) return;
+    if (pe == null) return;
     UIHelper.showConfirmDialog(getActivity(),
-        (getString(R.string.delete_public_holiday) + " '" + pe.getName() + "'" + getString(R.string.help)),
-        (v) -> {
-          mAdapter.removeItem(pe);
-          mApp.getProfilesFactory().remove(pe);
-          UIHelper.snack(mActivity, getString(R.string.profile_removed),
-              getString(R.string.undo), (nullview) -> {
-                mAdapter.addItem(pe);
-                mApp.getProfilesFactory().add(pe);
-              });
-        });
+      (getString(R.string.delete_profile) + " '" + pe.getName() + "'" + getString(R.string.help)),
+      v -> {
+        mAdapter.removeItem(pe);
+        mApp.getProfilesFactory().remove(pe);
+        UIHelper.snack(mActivity, getString(R.string.profile_removed),
+          getString(R.string.undo), nullview -> {
+            mAdapter.addItem(pe);
+            mApp.getProfilesFactory().add(pe);
+          });
+      });
 
   }
 }
