@@ -1,14 +1,12 @@
 package fr.ralala.worktime.ui.fragments;
 
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.res.Configuration;
 import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AlertDialog;
 import android.util.DisplayMetrics;
 import android.util.Log;
 import android.util.SparseArray;
@@ -20,6 +18,10 @@ import android.widget.LinearLayout;
 import android.widget.TableLayout;
 import android.widget.TableRow;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.fragment.app.Fragment;
 
 import org.achartengine.GraphicalView;
 import org.achartengine.chart.BarChart;
@@ -44,21 +46,22 @@ import java.util.TreeSet;
 
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
-import fr.ralala.worktime.ui.activities.MainActivity;
 import fr.ralala.worktime.models.DayEntry;
 import fr.ralala.worktime.models.DayType;
 import fr.ralala.worktime.models.WorkTimeDay;
-import fr.ralala.worktime.utils.AndroidHelper;
+import fr.ralala.worktime.ui.activities.MainActivity;
 import fr.ralala.worktime.ui.utils.UIHelper;
+import fr.ralala.worktime.utils.AndroidHelper;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Management of the charts fragment view
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
 public class StatisticsFragment extends Fragment implements View.OnClickListener {
   private static final int BACK_TIME_DELAY = 2000;
@@ -87,7 +90,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
   private Button btAction;
   private SparseArray<SummaryEntry> mSummaries;
 
-  private enum CurrentView{
+  private enum CurrentView {
     YEARS,
     MONTHS,
     WEEKS,
@@ -95,7 +98,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     DETAIL
   }
 
-  private class SummaryEntry {
+  private static class SummaryEntry {
     double nbWork;
     double nbHolidays;
     double nbSickness;
@@ -105,20 +108,20 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     long hWorkOver;
   }
 
-  private class ChartEntry {
+  private static class ChartEntry {
     WorkTimeDay work = new WorkTimeDay();
     WorkTimeDay over = new WorkTimeDay();
     WorkTimeDay pause = new WorkTimeDay();
     WorkTimeDay morning = new WorkTimeDay();
     WorkTimeDay afternoon = new WorkTimeDay();
-    DayType dt_morning = DayType.ERROR;
-    DayType dt_afternoon = DayType.ERROR;
-    String x_label;
+    DayType dtMorning = DayType.ERROR;
+    DayType dtAfternoon = DayType.ERROR;
+    String xLabel;
     double dover;
     double dwork;
   }
 
-  private class AnnotationEntry {
+  private static class AnnotationEntry {
     int idx;
     String annotation;
     double x;
@@ -134,10 +137,11 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Called when backPressed is consumed.
+   *
    * @return boolean
    */
   public boolean consumeBackPressed() {
-    if(mCurrentView == CurrentView.YEARS) {
+    if (mCurrentView == CurrentView.YEARS) {
       if (mLastBackPressed + BACK_TIME_DELAY > System.currentTimeMillis()) {
         return false;
       } else {
@@ -145,14 +149,13 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
         AndroidHelper.vibrate(mActivity);
       }
       mLastBackPressed = System.currentTimeMillis();
-    }
-    else if(mCurrentView == CurrentView.DETAIL)
+    } else if (mCurrentView == CurrentView.DETAIL)
       mCurrentView = CurrentView.DAYS;
-    else if(mCurrentView == CurrentView.DAYS)
+    else if (mCurrentView == CurrentView.DAYS)
       mCurrentView = CurrentView.WEEKS;
-    else if(mCurrentView == CurrentView.WEEKS)
+    else if (mCurrentView == CurrentView.WEEKS)
       mCurrentView = CurrentView.MONTHS;
-    else if(mCurrentView == CurrentView.MONTHS)
+    else if (mCurrentView == CurrentView.MONTHS)
       mCurrentView = CurrentView.YEARS;
     redrawChart();
     return true;
@@ -160,68 +163,73 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Adds a day entry.
+   *
    * @param entries The entries list.
-   * @param key The list key.
-   * @param de The current day entry.
-   * @param x_label The x label.
+   * @param key     The list key.
+   * @param de      The current day entry.
+   * @param xLabel  The x label.
    */
-  private void addDayEntry(SparseArray<ChartEntry> entries, Integer key, DayEntry de, String x_label) {
+  private void addDayEntry(SparseArray<ChartEntry> entries, Integer key, DayEntry de, String xLabel) {
     if (AndroidHelper.notContainsKey(entries, key)) {
       entries.put(key, new ChartEntry());
     }
     entries.get(key).work.addTime(de.getWorkTime());
-    if(!(de.getTypeAfternoon() == DayType.HOLIDAY  && de.getTypeMorning() == DayType.HOLIDAY) && !(de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY  && de.getTypeMorning() == DayType.PUBLIC_HOLIDAY))
+    if (!(de.getTypeAfternoon() == DayType.HOLIDAY && de.getTypeMorning() == DayType.HOLIDAY) && !(de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY && de.getTypeMorning() == DayType.PUBLIC_HOLIDAY))
       entries.get(key).over.addTime(de.getOverTime());
-    entries.get(key).x_label = x_label;
-    WorkTimeDay wm = de.isValidMorningType() ? de.getEndMorning().clone() : new WorkTimeDay();
-    if(!wm.timeString().equals("00:00"))
-      wm.delTime(de.isValidMorningType() ? de.getStartMorning().clone() : new WorkTimeDay());
+    entries.get(key).xLabel = xLabel;
+    WorkTimeDay wm = de.isValidMorningType() ? de.getEndMorning().copy() : new WorkTimeDay();
+    if (!wm.timeString().equals("00:00"))
+      wm.delTime(de.isValidMorningType() ? de.getStartMorning().copy() : new WorkTimeDay());
     entries.get(key).morning = wm;
-    WorkTimeDay wa = de.isValidAfternoonType() ? de.getEndAfternoon().clone() : new WorkTimeDay();
-    if(!wa.timeString().equals("00:00"))
-      wa.delTime(de.isValidAfternoonType() ? de.getStartAfternoon().clone() : new WorkTimeDay());
+    WorkTimeDay wa = de.isValidAfternoonType() ? de.getEndAfternoon().copy() : new WorkTimeDay();
+    if (!wa.timeString().equals("00:00"))
+      wa.delTime(de.isValidAfternoonType() ? de.getStartAfternoon().copy() : new WorkTimeDay());
     entries.get(key).afternoon = wa;
     entries.get(key).pause = de.getPause();
-    entries.get(key).dt_afternoon = de.getTypeAfternoon();
-    entries.get(key).dt_morning = de.getTypeMorning();
+    entries.get(key).dtAfternoon = de.getTypeAfternoon();
+    entries.get(key).dtMorning = de.getTypeMorning();
 
-    if(mSummaries != null) {
+    if (mSummaries != null) {
       if (AndroidHelper.notContainsKey(mSummaries, key)) {
         mSummaries.put(key, new SummaryEntry());
       }
       SummaryEntry se = mSummaries.get(key);
       switch (de.getTypeMorning()) {
         case AT_WORK:
-          se.nbWork+=0.5;
+          se.nbWork += 0.5;
           break;
         case HOLIDAY:
-          se.nbHolidays+=0.5;
+          se.nbHolidays += 0.5;
           break;
         case SICKNESS:
-          se.nbSickness+=0.5;
+          se.nbSickness += 0.5;
           break;
         case UNPAID:
-          se.nbUnpaid+=0.5;
+          se.nbUnpaid += 0.5;
           break;
         case RECOVERY:
-          se.nbRecovery+=0.5;
+          se.nbRecovery += 0.5;
+          break;
+        default:
           break;
       }
       switch (de.getTypeAfternoon()) {
         case AT_WORK:
-          se.nbWork+=0.5;
+          se.nbWork += 0.5;
           break;
         case HOLIDAY:
-          se.nbHolidays+=0.5;
+          se.nbHolidays += 0.5;
           break;
         case SICKNESS:
-          se.nbSickness+=0.5;
+          se.nbSickness += 0.5;
           break;
         case UNPAID:
-          se.nbUnpaid+=0.5;
+          se.nbUnpaid += 0.5;
           break;
         case RECOVERY:
-          se.nbRecovery+=0.5;
+          se.nbRecovery += 0.5;
+          break;
+        default:
           break;
       }
       se.hWork = entries.get(key).work.getTimeMs();
@@ -231,8 +239,9 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Called when the fragment is created.
-   * @param inflater The fragment inflater.
-   * @param container The fragment container.
+   *
+   * @param inflater           The fragment inflater.
+   * @param container          The fragment container.
    * @param savedInstanceState The saved instance state.
    * @return The created view.
    */
@@ -240,16 +249,16 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
   public View onCreateView(@NonNull final LayoutInflater inflater,
                            final ViewGroup container, final Bundle savedInstanceState) {
     final ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.fragment_statistics, container, false);
-    mActivity = (MainActivity)getActivity();
+    mActivity = (MainActivity) getActivity();
     assert mActivity != null;
-    mApp = MainApplication.getInstance();
+    mApp = (MainApplication) mActivity.getApplication();
     btAction = rootView.findViewById(R.id.action);
     btAction.setOnClickListener(this);
     mChartContainer = rootView.findViewById(R.id.graph);
     mMetrics = new DisplayMetrics();
     mActivity.getWindowManager().getDefaultDisplay().getMetrics(mMetrics);
     /* uses 80% of the screen */
-    mChartContainer.getLayoutParams().height = (int)getPercentOf(mMetrics.heightPixels, 85);
+    mChartContainer.getLayoutParams().height = (int) getPercentOf(mMetrics.heightPixels, 85);
     mDefaultColor = getResources().getColor(R.color.half_black, mActivity.getTheme());
     mColorUnder = getResources().getColor(R.color.color_under, mActivity.getTheme());
     mColorOver = getResources().getColor(R.color.color_over, mActivity.getTheme());
@@ -266,13 +275,14 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Called when the user click on a bar.
+   *
    * @param view The current view.
    */
   @Override
   public void onClick(View view) {
-    if(view.equals(btAction)) {
-      if(btAction.getText().toString().equals(getString(R.string.statistics_summary))) {
-        if(mSummaries == null) {
+    if (view.equals(btAction)) {
+      if (btAction.getText().toString().equals(getString(R.string.statistics_summary))) {
+        if (mSummaries == null) {
           UIHelper.showAlertDialog(mActivity, R.string.statistics_summary, getString(R.string.statistics_summary_empty));
           return;
         }
@@ -286,10 +296,11 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Called when the configuration is changed.
+   *
    * @param newConfig The new configuration.
    */
   @Override
-  public void onConfigurationChanged(Configuration newConfig) {
+  public void onConfigurationChanged(@NonNull Configuration newConfig) {
     super.onConfigurationChanged(newConfig);
     redrawChart();
   }
@@ -300,58 +311,59 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
   public void redrawChart() {
     String title = null;
 
-    if(mCurrentView == CurrentView.YEARS) {
+    if (mCurrentView == CurrentView.YEARS) {
       btAction.setCompoundDrawablesWithIntrinsicBounds(R.mipmap.ic_summary, 0, 0, 0);
       btAction.setText(R.string.statistics_summary);
-      if(mSummaries != null)
+      if (mSummaries != null)
         mSummaries.clear();
       mSummaries = new SparseArray<>();
     } else {
       btAction.setCompoundDrawablesWithIntrinsicBounds(R.mipmap.ic_cancel_white, 0, 0, 0);
       btAction.setText(R.string.statistics_reload);
-      if(mSummaries != null)
+      if (mSummaries != null)
         mSummaries.clear();
       mSummaries = null;
     }
     SparseArray<ChartEntry> entries = new SparseArray<>();
     boolean addPublicHoliday = true;
     List<Integer> years = mApp.getDaysFactory().getYears();
-    for(int nYear = 0; nYear < years.size(); nYear++) {
+    for (int nYear = 0; nYear < years.size(); nYear++) {
       Integer year = years.get(nYear);
-      if(mCurrentView == CurrentView.YEARS || mCurrentYearIndex == nYear) {
+      if (mCurrentView == CurrentView.YEARS || mCurrentYearIndex == nYear) {
         if (mCurrentView != CurrentView.YEARS) title = "" + year;
         List<Integer> months = mApp.getDaysFactory().getMonths(year);
-        for(int nMonth = 0; nMonth < months.size(); nMonth++) {
+        for (int nMonth = 0; nMonth < months.size(); nMonth++) {
           Integer month = months.get(nMonth);
-          if(mCurrentView == CurrentView.YEARS || mCurrentView == CurrentView.MONTHS || mCurrentMonthIndex == nMonth) {
+          if (mCurrentView == CurrentView.YEARS || mCurrentView == CurrentView.MONTHS || mCurrentMonthIndex == nMonth) {
             Map<Integer, List<DayEntry>> weeks = mApp.getDaysFactory().getWeeksAndDays(year, month);
             SortedSet<Integer> keysWeeks = new TreeSet<>(weeks.keySet());
             int nWeek = 0;
             for (Integer week : keysWeeks) {
-              if(mCurrentView == CurrentView.YEARS || mCurrentView == CurrentView.MONTHS || mCurrentView == CurrentView.WEEKS || mCurrentWeekIndex == nWeek) {
+              if (mCurrentView == CurrentView.YEARS || mCurrentView == CurrentView.MONTHS || mCurrentView == CurrentView.WEEKS || mCurrentWeekIndex == nWeek) {
                 /* Includes public holiday for the days view */
-                if(addPublicHoliday && mCurrentView == CurrentView.DAYS) {
+                if (addPublicHoliday && mCurrentView == CurrentView.DAYS) {
                   addPublicHoliday = false;
-                  mApp.getPublicHolidaysFactory().list(-1, month).forEach((ph) -> {
-                    if(ph.getDay().getYear() == year || (ph.getDay().getYear() != year && ph.isRecurrence())) {
+                  mApp.getPublicHolidaysFactory().list(-1, month).forEach(ph -> {
+                    if (ph.getDay().getYear() == year || (ph.getDay().getYear() != year && ph.isRecurrence())) {
                       DayEntry dePh = ph.toDayEntry();
                       dePh.getDay().setYear(year);
                       int w = dePh.getDay().toCalendar().get(Calendar.WEEK_OF_YEAR);
-                      if(w != week)
+                      if (w != week)
                         return;
-                      Integer e = dePh.getDay().toCalendar().get(Calendar.DAY_OF_WEEK) - 1;
-                      String x_label = getResources().getStringArray(R.array.days_letter)[e - 1] + " " + dePh.getDay().getDay();
-                      addDayEntry(entries, e, dePh, x_label);
+                      int e = dePh.getDay().toCalendar().get(Calendar.DAY_OF_WEEK) - 1;
+                      String xLabel = getResources().getStringArray(R.array.days_letter)[e - 1] + " " + dePh.getDay().getDay();
+                      addDayEntry(entries, e, dePh, xLabel);
                     }
                   });
                 }
                 List<DayEntry> days = weeks.get(week);
-                for (int nDay = 0; nDay < days.size(); nDay++) {
-                  DayEntry de = days.get(nDay);
-                  String s = parseDay(de, nDay, year, month, week, entries);
-                  if(s != null)
-                    title = s;
-                }
+                if (days != null)
+                  for (int nDay = 0; nDay < days.size(); nDay++) {
+                    DayEntry de = days.get(nDay);
+                    String s = parseDay(de, nDay, year, month, week, entries);
+                    if (s != null)
+                      title = s;
+                  }
               }
               nWeek++;
             }
@@ -359,7 +371,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
         }
       }
     }
-    if((mCurrentView != CurrentView.DETAIL))
+    if ((mCurrentView != CurrentView.DETAIL))
       drawBarChart(mMetrics.widthPixels, title, getString(R.string.statistics_hours), entries);
     else
       drawPieChart(title, entries);
@@ -368,11 +380,12 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Parse a day an adds the entry to the list.
-   * @param de The current DayEntry.
-   * @param nDay The current day index.
-   * @param year The current year.
-   * @param month The current month.
-   * @param week The current week.
+   *
+   * @param de      The current DayEntry.
+   * @param nDay    The current day index.
+   * @param year    The current year.
+   * @param month   The current month.
+   * @param week    The current week.
    * @param entries The entries list.
    * @return The title (or null).
    */
@@ -380,32 +393,30 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     String title = null;
     int day = de.getDay().getDay();
     if (mCurrentView == CurrentView.YEARS || mCurrentView == CurrentView.MONTHS || mCurrentView == CurrentView.WEEKS || mCurrentView == CurrentView.DAYS || mCurrentDayIndex == nDay) {
-      //if ((de.getTypeMorning() == DayType.PUBLIC_HOLIDAY && de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY) || (de.getTypeMorning() == DayType.HOLIDAY && de.getTypeAfternoon() == DayType.HOLIDAY))
-      //  return null;
       if (!de.isValidMorningType() && !de.isValidAfternoonType())
         return null;
-      Integer e = 0;
-      String x_label = "";
+      int e = 0;
+      String xLabel = "";
       if (mCurrentView == CurrentView.YEARS) {
         e = year;
-        x_label = "" + e;
+        xLabel = "" + e;
       } else if (mCurrentView == CurrentView.MONTHS) {
         e = month;
-        x_label = getResources().getStringArray(R.array.month_letter)[e - 1];
+        xLabel = getResources().getStringArray(R.array.month_letter)[e - 1];
         title = String.format(Locale.US, "%d", year);
       } else if (mCurrentView == CurrentView.WEEKS) {
         e = week;
-        x_label = getString(R.string.week_letter) + " " + e;
+        xLabel = getString(R.string.week_letter) + " " + e;
         title = String.format(Locale.US, "%d %s", year, getResources().getStringArray(R.array.month_short)[month - 1]);
       } else if (mCurrentView == CurrentView.DAYS) {
         e = de.getDay().toCalendar().get(Calendar.DAY_OF_WEEK) - 1;
-        x_label = getResources().getStringArray(R.array.days_letter)[e - 1] + " " + de.getDay().getDay();
+        xLabel = getResources().getStringArray(R.array.days_letter)[e - 1] + " " + de.getDay().getDay();
         title = String.format(Locale.US, "%d %s %s %d", year, getResources().getStringArray(R.array.month_short)[month - 1], getString(R.string.week), week);
       } else if (mCurrentView == CurrentView.DETAIL) {
         e = day;
         title = String.format(Locale.US, "%d %s %d (%s %d)", day, getResources().getStringArray(R.array.month_short)[month - 1], year, getString(R.string.week), week);
       }
-      addDayEntry(entries, e, de, x_label);
+      addDayEntry(entries, e, de, xLabel);
     }
     return title;
   }
@@ -421,35 +432,38 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Converts the time to a double value.
+   *
    * @param w The time to convert.
    * @return double
    */
   private double toDouble(WorkTimeDay w) {
-    return Double.parseDouble(w.timeString().replaceAll(":", "\\."));
+    return Double.parseDouble(w.timeString().replace(":", "."));
   }
 
   /**
    * Converts a double to a time string.
+   *
    * @param d The double to convert.
    * @return String
    */
   private String fromDouble(double d) {
     DecimalFormat df = new DecimalFormat("#00.00");
-    return df.format(d).replaceAll(",", ":");
+    return df.format(d).replace(",", ":");
   }
 
   /**
    * Initializes the multi render.
-   * @param title Milti render title.
+   *
+   * @param title       Milti render title.
    * @param legendHours Legend used for the hours.
    */
   private void initMultiRender(final String title, final String legendHours) {
-    if(mMultiRenderer != null) {
+    if (mMultiRenderer != null) {
       mMultiRenderer.clearXTextLabels();
       mMultiRenderer.clearYTextLabels();
       mMultiRenderer.removeAllRenderers();
     }
-    if(mChartContainer != null)
+    if (mChartContainer != null)
       mChartContainer.removeAllViews();
 
     mMultiRenderer = new XYMultipleSeriesRenderer();
@@ -473,7 +487,6 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     mMultiRenderer.setShowLegend(false);
     mMultiRenderer.setLegendHeight(30);
     mMultiRenderer.setXLabelsAlign(Paint.Align.CENTER);
-    //mMultiRenderer.setYLabelsAlign(Paint.Align.LEFT);
 
     mMultiRenderer.setYAxisAlign(Paint.Align.LEFT, 0);
     mMultiRenderer.setYLabelsAlign(Paint.Align.LEFT, 0);
@@ -494,7 +507,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     //setting the margin size for the graph in the order top, left, bottom, right
     mMultiRenderer.setMargins(new int[]{30, 30, 30, 30});
 
-    if(mCurrentView != CurrentView.DETAIL) {
+    if (mCurrentView != CurrentView.DETAIL) {
       mMultiRenderer.setChartTitle("");
       mMultiRenderer.setXTitle(title);
     } else {
@@ -507,13 +520,14 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Draw the bars chart.
-   * @param width Chart width.
-   * @param title Chart title.
+   *
+   * @param width       Chart width.
+   * @param title       Chart title.
    * @param legendHours Legend for the hours.
-   * @param values Bars value.
+   * @param values      Bars value.
    */
   private void drawBarChart(final int width, final String title, final String legendHours, final @NonNull SparseArray<ChartEntry> values) {
-    if(mDatasetBar != null)
+    if (mDatasetBar != null)
       mDatasetBar.clear();
     initMultiRender(title, legendHours);
 
@@ -524,14 +538,15 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     mMultiRenderer.setXAxisMax(keys.size() + 1);
     mMultiRenderer.setBarWidth((int) getPercentOf(width / mMultiRenderer.getXAxisMax(), 35));
     int i = 1;
-    double max = 0.0, min = Double.MAX_VALUE;
+    double max = 0.0;
+    double min = Double.MAX_VALUE;
 
     int seriesIdx = 0;
     List<AnnotationEntry> annotations = new ArrayList<>();
     for (Integer key : keys) {
       ChartEntry ce = values.get(key);
       WorkTimeDay withOver = ce.work; /* time include over */
-      WorkTimeDay work = withOver.clone();
+      WorkTimeDay work = withOver.copy();
       work.delTime(ce.over);
       ce.dover = toDouble(withOver);
       ce.dwork = toDouble(work);
@@ -547,7 +562,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
         addBarChartEntry(mDefaultColor, mColorOver, key, i, ce.dover);
         seriesIdx++;
         /* work */
-        if(mCurrentView == CurrentView.DAYS && ce.dt_morning == DayType.HOLIDAY && ce.dt_afternoon == DayType.HOLIDAY)
+        if (mCurrentView == CurrentView.DAYS && ce.dtMorning == DayType.HOLIDAY && ce.dtAfternoon == DayType.HOLIDAY)
           addBarChartEntry(mDefaultColor, mColorHoliday, key, i, ce.dwork);
         else
           addBarChartEntry(mDefaultColor, mColorWork, key, i, ce.dwork);
@@ -555,10 +570,10 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
         seriesIdx++;
       } else {
         String annotation;
-        if(mCurrentView == CurrentView.DAYS && !ce.work.isValidTime() && ce.dwork > 0 &&
-                ((ce.dt_morning == DayType.PUBLIC_HOLIDAY && ce.dt_afternoon == DayType.PUBLIC_HOLIDAY) || ce.dt_morning == DayType.HOLIDAY && ce.dt_afternoon == DayType.HOLIDAY)) {
-          int decimal = (int)ce.dwork;
-          int fractional = (int)Math.round((ce.dwork - decimal) * 100);
+        if (mCurrentView == CurrentView.DAYS && !ce.work.isValidTime() && ce.dwork > 0 &&
+          ((ce.dtMorning == DayType.PUBLIC_HOLIDAY && ce.dtAfternoon == DayType.PUBLIC_HOLIDAY) || ce.dtMorning == DayType.HOLIDAY && ce.dtAfternoon == DayType.HOLIDAY)) {
+          int decimal = (int) ce.dwork;
+          int fractional = (int) Math.round((ce.dwork - decimal) * 100);
           annotation = String.format(Locale.US, "%02d:%02d", decimal, fractional);
           addBarChartEntry(mDefaultColor, mColorPublicHoliday, key, i, ce.dwork);
           annotations.add(new AnnotationEntry(seriesIdx == 0 ? 0 : seriesIdx - 1, annotation, i, ce.dwork));
@@ -575,7 +590,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
         }
       }
 
-      mMultiRenderer.addXTextLabel(i, ce.x_label);
+      mMultiRenderer.addXTextLabel(i, ce.xLabel);
       if (ce.dover >= ce.dwork) {
         if (max < ce.dover) max = ce.dover;
         if (min > ce.dwork) min = ce.dwork;
@@ -604,7 +619,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     mChartView = new GraphicalView(mActivity, bchart);
 
     mMultiRenderer.setSelectableBuffer(100);
-    mChartView.setOnClickListener((v) -> {
+    mChartView.setOnClickListener(v -> {
       // handle the click event on the chart
       SeriesSelection seriesSelection = mChartView.getCurrentSeriesAndPoint();
       if (seriesSelection == null) {
@@ -633,31 +648,32 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Draw a pie chart.
-   * @param title Chart title.
+   *
+   * @param title  Chart title.
    * @param values Pie values.
    */
   private void drawPieChart(final String title, final @NonNull SparseArray<ChartEntry> values) {
     initMultiRender(title, null);
     final CategorySeries dataset = new CategorySeries("title");
     SortedSet<Integer> keys = toSortedSet(values);
-    final String [] titles = new String[3];
+    final String[] titles = new String[3];
     for (Integer key : keys) {
       int i = 0;
       ChartEntry ce = values.get(key);
       titles[i++] = getString(R.string.statistics_morning);
       titles[i++] = getString(R.string.statistics_break);
       titles[i] = getString(R.string.statistics_afternoon);
-      if(ce.dt_morning != DayType.ERROR &&  ce.dt_morning != DayType.RECOVERY)
+      if (ce.dtMorning != DayType.ERROR && ce.dtMorning != DayType.RECOVERY)
         addPieChartEntry(dataset, mColorMorning, getString(R.string.statistics_morning), toDouble(ce.morning));
-      if(ce.pause.isValidTime())
+      if (ce.pause.isValidTime())
         addPieChartEntry(dataset, mColorPause, getString(R.string.statistics_break), toDouble(ce.pause));
-      if(ce.dt_afternoon != DayType.ERROR &&  ce.dt_afternoon != DayType.RECOVERY)
+      if (ce.dtAfternoon != DayType.ERROR && ce.dtAfternoon != DayType.RECOVERY)
         addPieChartEntry(dataset, mColorAfternoon, getString(R.string.statistics_afternoon), toDouble(ce.afternoon));
     }
     PieChart pchart = new PieChart(dataset, mMultiRenderer);
     mChartView = new GraphicalView(mActivity, pchart);
     mMultiRenderer.setSelectableBuffer(3);
-    mChartView.setOnClickListener((v) -> {
+    mChartView.setOnClickListener(v -> {
       // handle the click event on the chart
       SeriesSelection seriesSelection = mChartView.getCurrentSeriesAndPoint();
       if (seriesSelection == null) {
@@ -666,7 +682,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
         int idx = seriesSelection.getPointIndex();
         SimpleSeriesRenderer ssr = mMultiRenderer.getSeriesRendererAt(idx);
         ssr.setHighlighted(!ssr.isHighlighted());
-        if(!ssr.isHighlighted())
+        if (!ssr.isHighlighted())
           dataset.set(idx, titles[idx], seriesSelection.getValue());
         else
           dataset.set(idx, fromDouble(seriesSelection.getValue()), seriesSelection.getValue());
@@ -678,7 +694,8 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Gets the percentage of a value.
-   * @param max Max value.
+   *
+   * @param max     Max value.
    * @param percent Required percent.
    * @return double.
    */
@@ -688,11 +705,12 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Adds a bar entry.
+   *
    * @param textColor The text color.
-   * @param inColor The bar color.
-   * @param key The XYSeries title.
-   * @param index The XYSeries index.
-   * @param value The XYSeries value.
+   * @param inColor   The bar color.
+   * @param key       The XYSeries title.
+   * @param index     The XYSeries index.
+   * @param value     The XYSeries value.
    */
   private void addBarChartEntry(int textColor, int inColor, int key, int index, double value) {
     XYSeriesRenderer renderer = new XYSeriesRenderer();
@@ -702,7 +720,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     renderer.setAnnotationsColor(textColor);
     renderer.setFillPoints(true);
     renderer.setDisplayChartValues(false);
-    XYSeries series = new XYSeries(""+key);
+    XYSeries series = new XYSeries("" + key);
     series.add(index, value);
     mDatasetBar.addSeries(series);
     mMultiRenderer.addSeriesRenderer(renderer);
@@ -710,13 +728,14 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Adds pie entry.
-   * @param dataset The owner dataset.
-   * @param inColor The background color.
+   *
+   * @param dataset  The owner dataset.
+   * @param inColor  The background color.
    * @param category The category name.
-   * @param value The category value.
+   * @param value    The category value.
    */
   private void addPieChartEntry(CategorySeries dataset, int inColor, String category, double value) {
-    SimpleSeriesRenderer  renderer = new SimpleSeriesRenderer();
+    SimpleSeriesRenderer renderer = new SimpleSeriesRenderer();
     renderer.setColor(inColor);
     renderer.setShowLegendItem(true);
     dataset.add(category, value);
@@ -728,24 +747,27 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
 
   /**
    * Converts entries keys to SortedSet
+   *
    * @param entries The entries.
    * @return SortedSet<Integer>
    */
   private SortedSet<Integer> toSortedSet(SparseArray<?> entries) {
     SortedSet<Integer> keys = new TreeSet<>();
-    for(int i = 0; i < entries.size(); i++)
+    for (int i = 0; i < entries.size(); i++)
       keys.add(entries.keyAt(i));
     return keys;
   }
 
   /**
    * Normalizes the value of a double avoiding the fractional part if it is equal to zero.
+   *
    * @param d The double to normalize.
    * @return String
    */
   private String fixDouble(double d) {
-    return (d % 1) == 0 ? String.valueOf((int)d) : String.valueOf(d);
+    return (d % 1) == 0 ? String.valueOf((int) d) : String.valueOf(d);
   }
+
   /**
    * Displays the summary alert dialog.
    */
@@ -764,7 +786,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* Year */
       TableRow tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       TextView textView = new TextView(dialogContext);
       textView.setTypeface(null, Typeface.BOLD);
       textView.setText(String.valueOf(year));
@@ -773,7 +795,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* hWork */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.work_time);
@@ -786,7 +808,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* hWorkOver */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.overtime);
@@ -799,7 +821,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* nbWork */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.at_work);
@@ -812,7 +834,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* nbHolidays */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.holidays);
@@ -825,7 +847,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* nbSickness */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.sickness);
@@ -838,7 +860,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* nbUnpaid */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.unpaid);
@@ -851,7 +873,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
       /* nbRecovery */
       tableRow = new TableRow(dialogContext);
       tableRow.setLayoutParams(new LinearLayout.LayoutParams
-          (LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT));
+        (ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
       tableRow.addView(new TextView(dialogContext)); /* empty view */
       textView = new TextView(dialogContext);
       textView.setText(R.string.recovery);
@@ -866,7 +888,7 @@ public class StatisticsFragment extends Fragment implements View.OnClickListener
     builder.setCancelable(true);
     AlertDialog alertDialog = builder.create();
     alertDialog.setTitle(getString(R.string.statistics_summary));
-    alertDialog.setButton(AlertDialog.BUTTON_POSITIVE, getString(R.string.ok), (dialog, which) -> dialog.dismiss());
+    alertDialog.setButton(DialogInterface.BUTTON_POSITIVE, getString(R.string.ok), (dialog, which) -> dialog.dismiss());
     alertDialog.show();
   }
 }
