@@ -1,9 +1,6 @@
 package fr.ralala.worktime.ui.fragments;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,37 +12,40 @@ import android.widget.ListView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
+import androidx.activity.result.ActivityResult;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 import java.util.Locale;
 
-import fr.ralala.worktime.models.PublicHolidayEntry;
-import fr.ralala.worktime.ui.activities.DayActivity;
-import fr.ralala.worktime.ui.activities.MainActivity;
-
-import fr.ralala.worktime.ui.dialogs.MonthDetailsDialog;
-import fr.ralala.worktime.services.QuickAccessService;
-import fr.ralala.worktime.utils.AndroidHelper;
 import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
-import fr.ralala.worktime.ui.adapters.DaysEntriesArrayAdapter;
+import fr.ralala.worktime.launchers.LauncherCallback;
 import fr.ralala.worktime.models.DayEntry;
 import fr.ralala.worktime.models.DayType;
+import fr.ralala.worktime.models.PublicHolidayEntry;
 import fr.ralala.worktime.models.WorkTimeDay;
+import fr.ralala.worktime.ui.activities.MainActivity;
+import fr.ralala.worktime.ui.adapters.DaysEntriesArrayAdapter;
+import fr.ralala.worktime.ui.dialogs.MonthDetailsDialog;
 import fr.ralala.worktime.ui.utils.SwipeDetector;
 import fr.ralala.worktime.ui.utils.UIHelper;
+import fr.ralala.worktime.utils.AndroidHelper;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Management of the main fragment view
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
-public class WorkTimeFragment extends Fragment implements View.OnClickListener, AdapterView.OnItemClickListener, SwipeDetector.SwipeDetectorListener {
+public class WorkTimeFragment extends Fragment implements View.OnClickListener, AdapterView.OnItemClickListener, SwipeDetector.SwipeDetectorListener, LauncherCallback {
 
   private ImageButton mBtPreviousMonth = null;
   private ImageButton mBtNextMonth = null;
@@ -62,8 +62,9 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
 
   /**
    * Called when the fragment is created.
-   * @param inflater The fragment inflater.
-   * @param container The fragment container.
+   *
+   * @param inflater           The fragment inflater.
+   * @param container          The fragment container.
    * @param savedInstanceState The saved instance state.
    * @return The created view.
    */
@@ -71,11 +72,10 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
   public View onCreateView(@NonNull final LayoutInflater inflater,
                            final ViewGroup container, final Bundle savedInstanceState) {
     final ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.fragment_main, container, false);
-    mActivity = (MainActivity)getActivity();
+    mActivity = (MainActivity) getActivity();
     assert mActivity != null;
     mActivity.getSwipeDetector().setSwipeDetectorListener(this);
-    mApp = MainApplication.getInstance();
-    //mApp.getCurrentDate().setTime(new Date());
+    mApp = (MainApplication) mActivity.getApplication();
 
     mMonthDetailsDialog = new MonthDetailsDialog(getActivity(), mApp);
 
@@ -98,28 +98,30 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
     mDays.setOnItemClickListener(this);
     LinearLayout llYearMonth = rootView.findViewById(R.id.llYearMonth);
     llYearMonth.setOnClickListener((v) ->
-        UIHelper.openDatePicker(getActivity(), mApp.getCurrentDate(), (view, selectedYear, selectedMonth, selectedDay) -> {
-          mApp.getCurrentDate().set(Calendar.YEAR, selectedYear);
-          mApp.getCurrentDate().set(Calendar.MONTH, selectedMonth);
-          mApp.getCurrentDate().set(Calendar.DAY_OF_MONTH, selectedDay);
+      UIHelper.openDatePicker(getActivity(), mApp.getCurrentDate(), (view, selectedYear, selectedMonth, selectedDay) -> {
+        mApp.getCurrentDate().set(Calendar.YEAR, selectedYear);
+        mApp.getCurrentDate().set(Calendar.MONTH, selectedMonth);
+        mApp.getCurrentDate().set(Calendar.DAY_OF_MONTH, selectedDay);
         updateAll();
       })
     );
 
-    mDays.setOnScrollListener(new AbsListView.OnScrollListener(){
+    mDays.setOnScrollListener(new AbsListView.OnScrollListener() {
       public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
+        /* Nothing */
       }
+
       public void onScrollStateChanged(AbsListView view, int scrollState) {
         if (view.getId() == mDays.getId()) {
           final int currentFirstVisibleItem = mDays.getFirstVisiblePosition();
           boolean isScrollingUp = (currentFirstVisibleItem < mApp.getLastFirstVisibleItem());
           /* change the visibility if 5% of the list is displayed or hidden */
           int k = at5Percent();
-          if(!isScrollingUp && currentFirstVisibleItem > k && mRlDetails.getVisibility() == View.VISIBLE)
+          if (!isScrollingUp && currentFirstVisibleItem > k && mRlDetails.getVisibility() == View.VISIBLE)
             mRlDetails.setVisibility(View.GONE);
-          else if(isScrollingUp && currentFirstVisibleItem < k && mRlDetails.getVisibility() == View.GONE)
+          else if (isScrollingUp && currentFirstVisibleItem < k && mRlDetails.getVisibility() == View.GONE)
             mRlDetails.setVisibility(View.VISIBLE);
-          else if(currentFirstVisibleItem == 0 && mRlDetails.getVisibility() != View.VISIBLE)
+          else if (currentFirstVisibleItem == 0 && mRlDetails.getVisibility() != View.VISIBLE)
             mRlDetails.setVisibility(View.VISIBLE);
           /* store previous item */
           mApp.setLastFirstVisibleItem(currentFirstVisibleItem);
@@ -130,58 +132,49 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
     return rootView;
   }
 
-  /**
-   * Receive the result from a previous call to startActivityForResult
-   * @param requestCode The integer request code originally supplied to startActivityForResult.
-   * @param resultCode The integer result code returned by the child activity through its setResult().
-   * @param data An Intent, which can return result data to the caller (various data can be attached to Intent "extras").
-   */
   @Override
-  public void onActivityResult(int requestCode, int resultCode, Intent data) {
-    if(requestCode == DayActivity.REQUEST_START_ACTIVITY)
-      updateAll();
+  public void onLauncherResult(ActivityResult result) {
+    updateAll();
   }
 
   /**
    * Called when the user click on a specific date.
+   *
    * @param adapterView The adapter view.
-   * @param view The view.
-   * @param i The view position.
-   * @param l See official javadoc (not used here).
+   * @param view        The view.
+   * @param i           The view position.
+   * @param l           See official javadoc (not used here).
    */
   @Override
   public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-    if(AndroidHelper.isServiceRunning(getActivity(), QuickAccessService.class)) {
-      UIHelper.toast(getActivity(), R.string.service_running);
-      return;
-    }
     DayEntry de = mLvAdapter.getItem(i);
-    if(de == null || de.getTypeMorning() == DayType.PUBLIC_HOLIDAY || de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY) {
+    if (de == null || de.getTypeMorning() == DayType.PUBLIC_HOLIDAY || de.getTypeAfternoon() == DayType.PUBLIC_HOLIDAY) {
       UIHelper.toast(getActivity(), R.string.error_editing_public_holiday);
       return;
     }
-    if(de.getWeekNumber() != DayEntry.INVALID_WEEK)
+    if (de.getWeekNumber() != DayEntry.INVALID_WEEK)
       return;
-    DayActivity.startActivity(this, de.getDay().dateString(), true);
+    mActivity.getLauncherDayActivity().startActivity(this, de.getDay().dateString(), true);
   }
 
   /**
    * Called when a button is clicked (previous, next and details).
+   *
    * @param v The clicked view.
    */
   public void onClick(final View v) {
-    if(v.equals(mBtPreviousMonth)) {
+    if (v.equals(mBtPreviousMonth)) {
       mApp.getCurrentDate().add(Calendar.MONTH, -1);
       mApp.setLastFirstVisibleItem(0);
       updateAll();
-    } else if(v.equals(mBtNextMonth)) {
+    } else if (v.equals(mBtNextMonth)) {
       mApp.getCurrentDate().add(Calendar.MONTH, 1);
       mApp.setLastFirstVisibleItem(0);
       updateAll();
-    } else if(v.equals(mRlDetails)) {
+    } else if (v.equals(mRlDetails)) {
       mMonthDetailsDialog.reloadDetails(
-          mApp.getCurrentDate().get(Calendar.MONTH),
-          mApp.getCurrentDate().get(Calendar.YEAR));
+        mApp.getCurrentDate().get(Calendar.MONTH),
+        mApp.getCurrentDate().get(Calendar.YEAR));
       mMonthDetailsDialog.open();
     }
   }
@@ -194,7 +187,7 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
     new Thread(() -> {
       updateTop();
       updateDates();
-       mActivity.runOnUiThread(() -> mActivity.progressDismiss());
+      mActivity.runOnUiThread(() -> mActivity.progressDismiss());
     }).start();
   }
 
@@ -218,10 +211,11 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
 
   /**
    * Returns 5% of the total visible items.
+   *
    * @return int
    */
   private int at5Percent() {
-    return (int)(mLvAdapter.getCount()*(5.0f/100.0f));
+    return (int) (mLvAdapter.getCount() * (5.0f / 100.0f));
   }
 
   /**
@@ -244,22 +238,22 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
     int weekNumber = DayEntry.INVALID_WEEK;
     List<DayEntry> dbDays = mApp.getDaysFactory().list(+mApp.getCurrentDate().get(Calendar.YEAR), +mApp.getCurrentDate().get(Calendar.MONTH) + 1, -1);
     /* loop for each days in the month */
-    for(int day = minDay; day <= maxDay; ++day) {
+    for (int day = minDay; day <= maxDay; ++day) {
       mApp.getCurrentDate().set(Calendar.DAY_OF_MONTH, day);
-      DayEntry de = new DayEntry(mApp.getCurrentDate(), DayType.ERROR, DayType.ERROR);
+      DayEntry de = new DayEntry(mActivity, mApp.getCurrentDate(), DayType.ERROR, DayType.ERROR);
 
-      if(mApp.isDisplayWeek() && weekNumber != mApp.getCurrentDate().get(Calendar.WEEK_OF_YEAR)) {
+      if (mApp.isDisplayWeek() && weekNumber != mApp.getCurrentDate().get(Calendar.WEEK_OF_YEAR)) {
         weekNumber = mApp.getCurrentDate().get(Calendar.WEEK_OF_YEAR);
         de.setWeekNumber(weekNumber);
         final DayEntry fde = de;
         mActivity.runOnUiThread(() -> mLvAdapter.add(fde));
-        de = new DayEntry(mApp.getCurrentDate(), DayType.ERROR, DayType.ERROR);
+        de = new DayEntry(mActivity, mApp.getCurrentDate(), DayType.ERROR, DayType.ERROR);
       }
 
       de.setAmountByHour(mApp.getAmountByHour()); /* set default amount */
       /* Force public holiday */
       boolean isPublicHoliday = mApp.getPublicHolidaysFactory().isPublicHolidays(publicHolidays, de.getDay());
-      if(isPublicHoliday) {
+      if (isPublicHoliday) {
         de.setTypeMorning(DayType.PUBLIC_HOLIDAY);
         de.setTypeAfternoon(DayType.PUBLIC_HOLIDAY);
       }
@@ -267,34 +261,35 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
       /* reload data if the current day is already inserted */
       mApp.getDaysFactory().checkForDayDateAndCopy(dbDays, de);
       /* count working day */
-      if(now != Calendar.SUNDAY && now != Calendar.SATURDAY && !isPublicHoliday) {
+      if (now != Calendar.SUNDAY && now != Calendar.SATURDAY && !isPublicHoliday) {
         wDays.add(de);
-        if(de.getTypeMorning() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.RECOVERY) realwDays += 0.5;
-        if(de.getTypeAfternoon() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.RECOVERY) realwDays += 0.5;
+        if (de.getTypeMorning() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.RECOVERY)
+          realwDays += 0.5;
+        if (de.getTypeAfternoon() == DayType.AT_WORK || de.getTypeAfternoon() == DayType.RECOVERY)
+          realwDays += 0.5;
       }
       final DayEntry fde = de;
       mActivity.runOnUiThread(() -> mLvAdapter.add(fde));
-      if(mApp.isScrollToCurrentDay() && mApp.getLastFirstVisibleItem() == 0 && de.getDay().dateString().equals(wtdnow.dateString())) {
+      if (mApp.isScrollToCurrentDay() && mApp.getLastFirstVisibleItem() == 0 && de.getDay().dateString().equals(wtdnow.dateString())) {
         mApp.setLastFirstVisibleItem(index);
-      }
-      else if(mApp.getLastFirstVisibleItem() == 0) index++;
+      } else if (mApp.getLastFirstVisibleItem() == 0) index++;
     }
 
 
     int min = (firstWeek == 52 ? 1 : firstWeek);
-    for(int w = min; w <= min + 6; ++w) {
-      WorkTimeDay wtdWorkTimeFromWeek =  mApp.getDaysFactory().getWorkTimeDayFromWeek(dbDays, w);
-      if(wtdWorkTimeFromWeek.isValidTime())
+    for (int w = min; w <= min + 6; ++w) {
+      WorkTimeDay wtdWorkTimeFromWeek = mApp.getDaysFactory().getWorkTimeDayFromWeek(dbDays, w);
+      if (wtdWorkTimeFromWeek.isValidTime())
         wtdTotalWorkTime.addTime(wtdWorkTimeFromWeek);
     }
     /* reload work day label */
     mApp.getCurrentDate().set(Calendar.DAY_OF_MONTH, currentDay);
-    int n = Integer.parseInt((""+realwDays).split("\\.")[1]);
+    int n = Integer.parseInt(("" + realwDays).split("\\.")[1]);
     String workDays = getString(R.string.work_days) + ": ";
-    if(n != 0)
-      workDays += String.format(Locale.US, "%02d.%02d/%02d", (int)realwDays, n, wDays.size());
+    if (n != 0)
+      workDays += String.format(Locale.US, "%02d.%02d/%02d", (int) realwDays, n, wDays.size());
     else
-      workDays += String.format(Locale.US, "%02d/%02d", (int)realwDays, wDays.size());
+      workDays += String.format(Locale.US, "%02d/%02d", (int) realwDays, wDays.size());
     workDays += " " + getString(R.string.days_lower_case);
     final String s_workDays = workDays;
     mActivity.runOnUiThread(() -> mTvWorkDays.setText(s_workDays));
@@ -309,7 +304,7 @@ public class WorkTimeFragment extends Fragment implements View.OnClickListener, 
     mActivity.runOnUiThread(() -> {
       mTvMonthlyHours.setText(monthlyHours);
       mLvAdapter.notifyDataSetChanged();
-    /* restores the scroll position and reloads the adapter else the listview seems not agree with the call of setSelection */
+      /* restores the scroll position and reloads the adapter else the listview seems not agree with the call of setSelection */
       mDays.setAdapter(mDays.getAdapter());
       mDays.setSelection(mApp.getLastFirstVisibleItem());
       if (mApp.isScrollToCurrentDay()) {
