@@ -6,23 +6,25 @@ import android.view.GestureDetector;
 import android.view.MotionEvent;
 
 /**
- *******************************************************************************
+ * ******************************************************************************
  * <p><b>Project WorkTime</b><br/>
  * Detect the swipe (left/right) gesture
  * </p>
- * @author Keidan
  *
- *******************************************************************************
+ * @author Keidan
+ * <p>
+ * ******************************************************************************
  */
-public class SwipeDetector implements GestureDetector.OnGestureListener{
+public class SwipeDetector implements GestureDetector.OnGestureListener {
   private static final int SWIPE_MIN_DISTANCE = 120;
   private static final int SWIPE_MAX_OFF_PATH = 250;
   private static final int SWIPE_THRESHOLD_VELOCITY = 200;
-  private GestureDetector mDetector;
+  private final GestureDetector mDetector;
   private SwipeDetectorListener mLi = null;
 
   /**
    * Creates the SwipeDetector instance.
+   *
    * @param c The Android context.
    */
   public SwipeDetector(final Context c) {
@@ -31,6 +33,7 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
 
   /**
    * Sets the swipe detector listener.
+   *
    * @param li The new listener.
    */
   public void setSwipeDetectorListener(final SwipeDetectorListener li) {
@@ -39,6 +42,7 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
 
   /**
    * Called in the activity when the onTouchEvent is called.
+   *
    * @param ev See official javadoc.
    * @return See official javadoc.
    */
@@ -48,6 +52,7 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
 
   /**
    * See official javadoc.
+   *
    * @param motionEvent See official javadoc.
    * @return true
    */
@@ -58,15 +63,17 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
 
   /**
    * See official javadoc.
+   *
    * @param motionEvent See official javadoc.
    */
   @Override
   public void onShowPress(MotionEvent motionEvent) {
-
+    /* Nothing */
   }
 
   /**
    * See official javadoc.
+   *
    * @param motionEvent See official javadoc.
    * @return true
    */
@@ -77,10 +84,11 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
 
   /**
    * See official javadoc.
-   * @param motionEvent See official javadoc.
+   *
+   * @param motionEvent  See official javadoc.
    * @param motionEvent1 See official javadoc.
-   * @param v See official javadoc.
-   * @param v1 See official javadoc.
+   * @param v            See official javadoc.
+   * @param v1           See official javadoc.
    * @return true
    */
   @Override
@@ -90,17 +98,19 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
 
   /**
    * See official javadoc.
+   *
    * @param motionEvent See official javadoc.
    */
   @Override
   public void onLongPress(MotionEvent motionEvent) {
-
+    /* Nothing */
   }
 
   /**
    * See official javadoc.
-   * @param start See official javadoc.
-   * @param finish See official javadoc.
+   *
+   * @param start     See official javadoc.
+   * @param finish    See official javadoc.
    * @param velocityX See official javadoc.
    * @param velocityY See official javadoc.
    * @return boolean
@@ -111,13 +121,11 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
     if (Math.abs(start.getY() - finish.getY()) > SWIPE_MAX_OFF_PATH)
       return false;
 
-    if(start.getX() - finish.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY) {
+    if (start.getX() - finish.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY) {
       // left to right swipe
-      if(mLi != null) mLi.leftToRightSwipe();
-    }
-    else if (finish.getX() - start.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY) {
-      // right to left swipe
-      if(mLi != null) mLi.rightToLeftSwipe();
+      if (mLi != null) mLi.leftToRightSwipe();
+    } else if (finish.getX() - start.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY && (mLi != null)) {
+      mLi.rightToLeftSwipe();
     }
     return true;
   }
@@ -127,6 +135,7 @@ public class SwipeDetector implements GestureDetector.OnGestureListener{
      * Called when the user sweeps from left to right.
      */
     void leftToRightSwipe();
+
     /**
      * Called when the user sweeps from right to left.
      */
